HPCC-18566 Add support for serializing ifblocks
