+++ ecl/hqlcpp/hqlsource.cpp
@@ -251,33 +251,6 @@ IHqlExpression * getVirtualReplacement(IHqlExpression * field, IHqlExpression * 
     return NULL;
 }
 
-static IHqlExpression * getExplicitlyPromotedCompare(IHqlExpression * filter)
-{
-    switch (filter->getOperator())
-    {
-    case no_in:
-    case no_notin:
-        return LINK(filter);
-    }
-    IHqlExpression * l = filter->queryChild(0);
-    IHqlExpression * r = filter->queryChild(1);
-    ITypeInfo * lType = queryUnqualifiedType(l->queryType());
-    ITypeInfo * rType = queryUnqualifiedType(r->queryType());
-    if (lType == rType)
-        return LINK(filter);
-
-    //Add explicit casts to the type.  ensureExprType won't add a (string) to a string2 field.
-    Owned<ITypeInfo> promotedType = getPromotedECLType(lType, rType);
-    HqlExprArray args;
-    if (lType == promotedType)
-        args.append(*LINK(l));
-    else
-        args.append(*createValue(no_implicitcast, LINK(promotedType), LINK(l)));
-    args.append(*ensureExprType(r, promotedType));
-    return filter->clone(args);
-}
-
-
 static IHqlExpression * createFileposCall(HqlCppTranslator & translator, IIdAtom * name, const char * provider, const char * rowname)
 {
     HqlExprArray args;
@@ -620,7 +593,7 @@ public:
         rawSteppingProject.extractFields(rawStepping);
     }
 
-    void checkKeyable(MonitorExtractor & monitors)
+    void checkKeyable(CppFilterExtractor & monitors)
     {
         if (!rawStepping.exists())
             return;
@@ -730,7 +703,7 @@ public:
     void buildKeyedLimitHelper(IHqlExpression * expr);
     void buildLimits(BuildCtx & classctx, IHqlExpression * expr, unique_id_t id);
     void buildReadMembers( IHqlExpression * expr);
-    void buildSteppingMeta(IHqlExpression * expr, MonitorExtractor * monitors);
+    void buildSteppingMeta(IHqlExpression * expr, CppFilterExtractor * monitors);
     void buildTransformBody(BuildCtx & transformCtx, IHqlExpression * expr, bool returnSize, bool ignoreFilters, bool bindInputRow);
     void checkDependencies(BuildCtx & ctx, IHqlExpression * expr);
     bool containsStepping(IHqlExpression * expr);
@@ -756,7 +729,7 @@ protected:
     void buildGroupAggregateCompareHelper(ParentExtract * extractBuilder, IHqlExpression * aggregate, HqlExprArray & recordFields, HqlExprArray & agrgegateFields);
     void buildGroupAggregateHashHelper(ParentExtract * extractBuilder, IHqlExpression * dataset, IHqlExpression * fields);
     void buildGroupAggregateProcessHelper(ParentExtract * extractBuilder, IHqlExpression * aggregate, const char * name, bool doneAny);
-    void buildGroupingMonitors(IHqlExpression * expr, MonitorExtractor & monitors);
+    void buildGroupingMonitors(IHqlExpression * expr, CppFilterExtractor & monitors);
     void buildGroupAggregateTransformBody(BuildCtx & transformctx, IHqlExpression * expr, bool useExtract, bool bindInputRow);
     void buildNormalizeHelpers(IHqlExpression * expr);
     void buildTargetCursor(Shared<BoundRow> & tempRow, Shared<BoundRow> & rowBuilder, BuildCtx & ctx, IHqlExpression * expr);
@@ -2415,7 +2388,7 @@ void SourceBuilder::buildGlobalGroupAggregateHelpers(IHqlExpression * expr)
 }
 
 
-void SourceBuilder::buildGroupingMonitors(IHqlExpression * expr, MonitorExtractor & monitors)
+void SourceBuilder::buildGroupingMonitors(IHqlExpression * expr, CppFilterExtractor & monitors)
 {
     IHqlExpression * aggregate = expr->queryChild(0);
     node_operator op = aggregate->getOperator();
@@ -2701,7 +2674,7 @@ class DiskReadBuilderBase : public SourceBuilder
 {
 public:
     DiskReadBuilderBase(HqlCppTranslator & _translator, IHqlExpression *_tableExpr, IHqlExpression *_nameExpr)
-        : SourceBuilder(_translator, _tableExpr, _nameExpr), monitors(_tableExpr, _translator, 0, true)
+        : SourceBuilder(_translator, _tableExpr, _nameExpr), monitors(_tableExpr, _translator, 0, true, false)
     {
         fpos.setown(getFilepos(tableExpr, false));
         lfpos.setown(getFilepos(tableExpr, true));
@@ -2719,7 +2692,7 @@ protected:
     virtual void buildFlagsMember(IHqlExpression * expr);
 
 protected:
-    MonitorExtractor monitors;
+    CppFilterExtractor monitors;
     OwnedHqlExpr globalGuard;
     IHqlExpression * mode;
     node_operator modeOp;
@@ -3612,2567 +3585,6 @@ ABoundActivity * HqlCppTranslator::doBuildActivityCompoundSelectNew(BuildCtx & c
     return info.buildActivity(ctx, activityExpr, TAKchildthroughnormalize, "ChildThroughNormalize", childActivity);
 }
 
-//-----------------------------------------------------------------------------------------------
-//-- Index processing
-//-----------------------------------------------------------------------------------------------
-
-/*
-
-  Note on generating segment monitors for conditions (ty)x = Y
-
-  1) casting from Tx to Ty loses information.  E.g., (int1)string2field = int1value.
-     In this case it is almost impossible to generate a segment monitor because we would need to work out all
-     the possibly values for x which could generate the value y.
-     The only exception is an inequality (and no_notin), which we can use to remove some of the candidates.  The test will always need
-     duplicating since we cannot remove all the expected values.
-
-  2) Casting from Ty to Tx loses information.  E.g., (string)string2field = stringvalue
-     In this case we can process the filter without prefiltering by testing whether isExact: (Ty)(Tx)Y == Y, and following the following rules:
-     a) no_eq.  If isExact, add value else nothing.
-     b) no_ne.  If isExtact, remove value else nothing.
-     c) no_gt.  Always add > value
-     d) no_ge.  If isExact add >= value else add > value
-     e) no_lt.  If isExact add < value else add <= value
-     f) no_le.  Always add <= value
-
-  3) Note casts must be present on both sides to indicate exactly what type the comparison will be done as.  This includes (string)which would
-     normally be missing if the field was of type string<n>.
-
-*/
-
-static node_operator getModifiedOp(node_operator op, bool duplicate)
-{
-    if (!duplicate)
-        return op;
-
-    switch (op)
-    {
-    case no_eq:
-    case no_le:
-    case no_ge:
-    case no_in:
-        return op;
-        //err on the side of caution for the segment monitors - 
-        //the test later which check it more accurately.
-    case no_gt:
-        return no_ge;
-    case no_lt:
-        return no_le;
-    case no_ne:
-    case no_notin:
-        return no_none;
-    default:
-        UNIMPLEMENTED;
-    }
-}
-
-
-void KeyFailureInfo::merge(const KeyFailureInfo & other)
-{
-    if (code < other.code)
-        set(other.code, other.field);
-}
-
-void KeyFailureInfo::reportError(HqlCppTranslator & translator, IHqlExpression * condition)
-{
-    StringBuffer ecl;
-    getExprECL(condition, ecl);
-    switch (code)
-    {
-    case KFRunknown:
-        translator.throwError1(HQLERR_KeyedJoinTooComplex, ecl.str());
-    case KFRnokey:
-        translator.throwError1(HQLERR_KeyAccessNoKeyField, ecl.str());
-    case KFRtoocomplex:
-        translator.throwError1(HQLERR_KeyedJoinTooComplex, ecl.str());
-    case KFRcast:
-        translator.throwError2(HQLERR_KeyAccessNeedCast, ecl.str(), str(field->queryName()));
-    case KFRor:
-        translator.throwError1(HQLERR_OrMultipleKeyfields, ecl.str());
-    }
-}
-
-
-const char * BuildMonitorState::getSetName(bool createValueSets)
-{
-    if (!setNames.isItem(numActiveSets))
-    {
-        StringBuffer name;
-        getUniqueId(name.append("set"));
-        setNames.append(*new StringAttrItem(name.str()));
-
-        StringBuffer s;
-        funcctx.setNextConstructor();
-        if (createValueSets)
-            funcctx.addQuoted(s.append("Owned<IValueSet> ").append(name).append(";"));
-        else
-            funcctx.addQuoted(s.append("Owned<IStringSet> ").append(name).append(";"));
-    }
-
-    return setNames.item(numActiveSets++).text;
-}
-
-
-void BuildMonitorState::popSetName()
-{
-    numActiveSets--;
-}
-
-
-IHqlExpression * KeyConditionInfo::createConjunction()
-{
-    LinkedHqlExpr result = preFilter;
-    ForEachItemIn(i, conditions)
-        extendAndCondition(result, conditions.item(i).expr);
-    extendAndCondition(result, postFilter);
-    return result.getClear();
-}
-
-//---------------------------------------------------------------------------------------------------------------------
-
-const char * KeySelectorInfo::getFFOptions()
-{
-    switch (keyedKind)
-    {
-    case KeyedExtend:
-        return "FFopt";
-    default:
-        return "FFkeyed";
-    }
-}
-
-//---------------------------------------------------------------------------------------------------------------------
-
-MonitorExtractor::MonitorExtractor(IHqlExpression * _tableExpr, HqlCppTranslator & _translator, int _numKeyableFields, bool _isDiskRead)
-    : translator(_translator), createValueSets(_isDiskRead || translator.queryOptions().createValueSets )
-{ 
-    tableExpr = _tableExpr;
-
-    if (_numKeyableFields <= 0)
-    {
-        //-ve number means remove a certain number of fields from the record
-        IHqlExpression * record = tableExpr->queryRecord();
-        numKeyableFields = 0;
-        ForEachChild(i, record)
-            if (!record->queryChild(i)->isAttribute())
-                numKeyableFields++;
-        numKeyableFields += _numKeyableFields;          // remove payload fields.
-    }
-    else
-        numKeyableFields = (unsigned)_numKeyableFields;
-
-    onlyHozedCompares = !_isDiskRead;
-
-    expandKeyableFields();
-    cleanlyKeyedExplicitly = false;
-    keyedExplicitly = false;
-    allowDynamicFormatChange = !tableExpr->hasAttribute(fixedAtom);
-    if (createValueSets)
-    {
-        addRangeFunc = addRawRangeId;
-        killRangeFunc = killRawRangeId;
-    }
-    else
-    {
-        addRangeFunc = addRangeId;
-        killRangeFunc = killRangeId;
-    }
-}
-
-void MonitorExtractor::callAddAll(BuildCtx & ctx, IHqlExpression * targetVar)
-{
-    HqlExprArray args;
-    args.append(*LINK(targetVar));
-    translator.callProcedure(ctx, addAllId, args);
-}
-
-static IHqlExpression * createExpandedRecord(IHqlExpression * expr);
-
-static ITypeInfo * getExpandedFieldType(ITypeInfo * type, IHqlExpression * expr)
-{
-    Linked<ITypeInfo> expandedType = type;
-    if (type->getSize() == UNKNOWN_LENGTH)
-        expandedType.clear();
-    switch (type->getTypeCode())
-    {
-    case type_packedint:
-        expandedType.setown(makeIntType(type->queryPromotedType()->getSize(), type->isSigned()));
-        break;
-    case type_bitfield:
-        expandedType.set(type->queryPromotedType());
-        break;
-    case type_varstring:
-    case type_varunicode:
-#if 0
-        if (type->getSize() != UNKNOWN_LENGTH)
-        {
-            unsigned len = type->getStringLen();
-            switch (type->getTypeCode())
-            {
-            case type_varstring:
-                expandedType.setown(makeStringType(len, LINK(type->queryCharset()), LINK(type->queryCollation())));
-                break;
-            case type_varunicode:
-                expandedType.setown(makeUnicodeType(len, type->queryLocale()));
-                break;
-            }
-            break;
-        }
-#endif              //fall through
-    case type_data:
-    case type_qstring:
-    case type_string:
-    case type_unicode:
-    case type_utf8:
-        if (type->getSize() == UNKNOWN_LENGTH)
-        {
-            unsigned maxLength = UNKNOWN_LENGTH;
-            IHqlExpression * maxSizeExpr = expr ? queryAttributeChild(expr, maxSizeAtom, 0) : NULL;
-            if (maxSizeExpr)
-            {
-                unsigned maxSize = (unsigned)maxSizeExpr->queryValue()->getIntValue();
-                switch (type->getTypeCode())
-                {
-                case type_data:
-                case type_string:
-                    maxLength = maxSize - sizeof(size32_t);
-                    break;
-                case type_qstring:
-                    maxLength = rtlQStrLength(maxSize - sizeof(size32_t));
-                    break;
-                case type_unicode:
-                    maxLength = (maxSize-sizeof(size32_t))/sizeof(UChar);
-                    break;
-                case type_utf8:
-                    maxLength = (maxSize-sizeof(size32_t))/4;
-                    break;
-                case type_varstring:
-                    maxLength = maxSize - 1;
-                    break;
-                case type_varunicode:
-                    maxLength = (maxSize/sizeof(UChar)) - 1;
-                    break;
-                }
-            }
-            else
-            {
-                IHqlExpression * maxLengthExpr = expr ? queryAttributeChild(expr, maxLengthAtom, 0) : NULL;
-                if (maxLengthExpr)
-                    maxLength = (unsigned)maxLengthExpr->queryValue()->getIntValue();
-            }
-
-            if (maxLength != UNKNOWN_LENGTH)
-            {
-                switch (type->getTypeCode())
-                {
-                case type_data:
-                    expandedType.setown(makeDataType(maxLength));
-                    break;
-                case type_qstring:
-                    expandedType.setown(makeQStringType(maxLength));
-                    break;
-                case type_string:
-                    expandedType.setown(makeStringType(maxLength, LINK(type->queryCharset()), LINK(type->queryCollation())));
-                    break;
-                case type_unicode:
-                    expandedType.setown(makeUnicodeType(maxLength, type->queryLocale()));
-                    break;
-                case type_utf8:
-                    expandedType.setown(makeUtf8Type(maxLength, type->queryLocale()));
-                    break;
-                case type_varstring:
-                    expandedType.setown(makeVarStringType(maxLength, LINK(type->queryCharset()), LINK(type->queryCollation())));
-                    break;
-                case type_varunicode:
-                    expandedType.setown(makeVarUnicodeType(maxLength, type->queryLocale()));
-                    break;
-                }
-            }
-        }
-        else
-        {
-            //This could ensure the strings are ascii, but the ebcdic strings are still comparable, and the order will be more logical
-            //if they remain as ebcdic.
-        }
-        break;
-    case type_table:
-    case type_groupedtable:
-    case type_set:
-        expandedType.clear();
-        break;
-    case type_row:
-        {
-            OwnedHqlExpr newRecord = createExpandedRecord(queryRecord(type));
-            if (isEmptyRecord(newRecord))
-                expandedType.clear();
-            else
-                expandedType.setown(makeRowType(LINK(newRecord->queryRecordType())));
-            break;
-        }
-    case type_alien:
-        {
-            IHqlAlienTypeInfo * alien = queryAlienType(type);
-            expandedType.set(alien->queryLogicalType());
-            break;
-        }
-    }
-    return expandedType.getClear();
-}
-
-static void createExpanded(HqlExprArray & fields, IHqlExpression * expr)
-{
-    switch (expr->getOperator())
-    {
-    case no_ifblock:
-        //if blocks need to generate translated segment monitors to be keyed, so don't expand them
-        break;
-    case no_record:
-        {
-            ForEachChild(i, expr)
-                createExpanded(fields, expr->queryChild(i));
-            break;
-        }
-    case no_field:
-        {
-            ITypeInfo * type = expr->queryType();
-            Owned<ITypeInfo> expandedType = getExpandedFieldType(type, expr);
-            if (expandedType)
-            {
-                if (expandedType == type)
-                    fields.append(*LINK(expr));
-                else
-                {
-                    HqlExprArray attrs;
-                    unwindChildren(attrs, expr);
-                    //MORE: Any default will now have the wrong type => remove it for the moment (ideally it would be projected)
-                    removeAttribute(attrs, defaultAtom);
-                    fields.append(*createField(expr->queryId(), LINK(expandedType), attrs));
-                }
-            }
-            break;
-        }
-    case no_attr:
-    case no_attr_link:
-    case no_attr_expr:
-        fields.append(*LINK(expr));
-        break;
-    }
-}
-
-static IHqlExpression * createExpandedRecord(IHqlExpression * expr)
-{
-    HqlExprArray fields;
-    createExpanded(fields, expr);
-    return cloneOrLink(expr, fields);
-}
-
-
-bool MonitorExtractor::createGroupingMonitor(BuildCtx ctx, const char * listName, IHqlExpression * expr, unsigned & maxField)
-{
-    switch (expr->getOperator())
-    {
-    case no_if:
-        {
-            IHqlExpression * cond = expr->queryChild(0);
-            if (expr->queryChild(2)->isConstant() && isIndependentOfScope(cond))
-            {
-                BuildCtx subctx(ctx);
-                translator.buildFilter(subctx, expr->queryChild(0));
-                createGroupingMonitor(subctx, listName, expr->queryChild(1), maxField);
-                return true;        // may still be keyed
-            }
-            break;
-        }
-    case no_select:
-        {
-            size32_t offset = 0;
-            ForEachItemIn(i, keyableSelects)
-            {
-                IHqlExpression & cur = keyableSelects.item(i);
-                size32_t curSize = cur.queryType()->getSize();
-                if (!createValueSets && curSize == UNKNOWN_LENGTH)
-                    break;
-                if (expr == &cur)
-                {
-                    maxField = i+1;
-                    if (createValueSets)
-                    {
-                        StringBuffer type;
-                        translator.buildRtlFieldType(type, expr->queryChild(1), queryRecord(tableExpr));
-                        ctx.addQuotedF("%s->append(FFkeyed, createWildFieldFilter(%u, %s));", listName, i, type.str());
-                    }
-                    else
-                    {
-                        //MORE: Check the type of the field is legal.
-                        ctx.addQuotedF("%s->append(createWildKeySegmentMonitor(%u, %u, %u));", listName, i, offset, curSize);
-                    }
-                    return true;
-                }
-                offset += curSize;
-            }
-            break;
-        }
-    case no_constant:
-        return true;
-    }
-    ctx.addReturn(queryBoolExpr(false));
-    return false;
-}
-
-void MonitorExtractor::expandKeyableFields()
-{
-    HqlExprArray fields;
-    IHqlExpression * tableRecord = tableExpr->queryRecord();
-    unsigned cnt = 0;
-    ForEachChild(i, tableRecord)
-    {
-        if (cnt == numKeyableFields)
-            break;
-        IHqlExpression * cur = tableRecord->queryChild(i);
-        if (!cur->isAttribute())
-        {
-            fields.append(*LINK(cur));
-            cnt++;
-        }
-    }
-    keyableRecord.setown(createRecord(fields));
-
-    if (createValueSets)
-        expandedRecord.set(keyableRecord);
-    else
-        expandedRecord.setown(createExpandedRecord(keyableRecord));
-    IHqlExpression * selector = tableExpr->queryNormalizedSelector();
-
-    OwnedHqlExpr expandedSelector = createDataset(no_anon, LINK(expandedRecord), createUniqueId());
-    firstOffsetField = NotFound;
-    expandSelects(keyableRecord, expandedRecord->querySimpleScope(), selector, expandedSelector);
-    if (firstOffsetField == NotFound)
-        firstOffsetField = keyableSelects.ordinality();
-}
-
-void MonitorExtractor::expandSelects(IHqlExpression * expr, IHqlSimpleScope * expandedScope, IHqlExpression * keySelector, IHqlExpression * expandedSelector)
-{
-    switch (expr->getOperator())
-    {
-    case no_record:
-        {
-            ForEachChild(i, expr)
-                expandSelects(expr->queryChild(i), expandedScope, keySelector, expandedSelector);
-            break;
-        }
-    case no_ifblock:
-        expandSelects(expr->queryChild(1), expandedScope, keySelector, expandedSelector);
-        break;
-    case no_field:
-        {
-            OwnedHqlExpr match = expandedScope->lookupSymbol(expr->queryId());
-            if (match)
-            {
-                OwnedHqlExpr keySelected = createSelectExpr(LINK(keySelector), LINK(expr));
-                OwnedHqlExpr expandedSelected = createSelectExpr(LINK(expandedSelector), LINK(match));
-                IHqlExpression * record = expr->queryRecord();
-                if (record)
-                    expandSelects(record, match->queryRecord()->querySimpleScope(), keySelected, expandedSelected);
-                else
-                {
-                    if ((expr != match) && (firstOffsetField == NotFound))
-                    {
-                        ITypeInfo * exprType = expr->queryType();
-                        ITypeInfo * matchType = match->queryType();
-                        if ((exprType->getSize() != matchType->getSize()) ||
-                            (exprType->getTypeCode() == type_bitfield || exprType->getTypeCode() == type_bitfield))
-                            firstOffsetField = keyableSelects.ordinality();
-                    }
-
-                    keyableSelects.append(*LINK(keySelected));
-                    expandedSelects.append(*LINK(expandedSelected));
-                }
-            }
-            else
-            {
-                if (firstOffsetField == NotFound)
-                    firstOffsetField = keyableSelects.ordinality();
-            }
-            break;
-        }
-    }
-}
-
-void MonitorExtractor::buildKeySegmentInExpr(BuildMonitorState & buildState, KeySelectorInfo & selectorInfo, BuildCtx & ctx, const char * target, IHqlExpression & thisKey, MonitorFilterKind filterKind)
-{
-    //Generally this slightly increases the code size, but reduces the number of
-    //temporary sets which is generally more efficient.
-    OwnedHqlExpr simplified = querySimplifyInExpr(&thisKey);
-    if (simplified)
-    {
-        OwnedHqlExpr folded = foldHqlExpression(simplified);
-        buildKeySegmentExpr(buildState, selectorInfo, ctx, target, *folded, filterKind);
-        return;
-    }
-
-    IHqlExpression * expandedSelector = selectorInfo.expandedSelector;
-    ITypeInfo * fieldType = expandedSelector->queryType();
-    unsigned curSize = fieldType->getSize();
-    createStringSet(ctx, target, curSize, expandedSelector);
-
-    OwnedHqlExpr targetVar = createVariable(target, makeVoidType());
-    IHqlExpression * lhs = thisKey.queryChild(0);
-    OwnedHqlExpr values = normalizeListCasts(thisKey.queryChild(1));
-
-    IIdAtom * func = addRangeFunc;
-    if (thisKey.getOperator() == no_notin)
-    {
-        callAddAll(ctx, targetVar);
-        func = killRangeFunc;
-    }
-
-    if (values->getOperator() != no_list)
-    {
-        //iterate through the set
-        BuildCtx subctx(ctx);
-        CHqlBoundExpr boundCurElement;
-        Owned<IHqlCppSetCursor> cursor = translator.createSetSelector(ctx, values);
-
-        bool done = false;
-        CHqlBoundExpr isAll;
-        cursor->buildIsAll(subctx, isAll);
-        if (isAll.expr->queryValue())
-        {
-            if (isAll.expr->queryValue()->getBoolValue())
-            {
-                callAddAll(subctx, targetVar);
-                done = true;
-                //If ALL allowed exceptions then we would need to do more....
-            }
-        }
-        else
-        {
-            IHqlStmt * stmt = subctx.addFilter(isAll.expr);
-            callAddAll(subctx, targetVar);
-            subctx.selectElse(stmt);
-        }
-
-        if (!done)
-        {
-            cursor->buildIterateLoop(subctx, boundCurElement, false);
-            OwnedHqlExpr curValue = boundCurElement.getTranslatedExpr();
-
-            OwnedHqlExpr test = createBoolExpr(no_eq, LINK(lhs), LINK(curValue));
-            OwnedHqlExpr promoted = getExplicitlyPromotedCompare(test);
-            OwnedHqlExpr compare, normalized;
-            extractCompareInformation(subctx, promoted, compare, normalized, expandedSelector);
-            if (compare)
-                translator.buildFilter(subctx, compare);
-
-            HqlExprArray args;
-            args.append(*LINK(targetVar));
-            unsigned srcSize = normalized->queryType()->getSize();
-            if (srcSize < curSize && curSize != UNKNOWN_LENGTH)
-            {
-                OwnedHqlExpr lengthExpr = getSizetConstant(srcSize);
-                OwnedHqlExpr rangeLower = getRangeLimit(fieldType, lengthExpr, normalized, -1);
-                OwnedHqlExpr rangeUpper = getRangeLimit(fieldType, lengthExpr, normalized, +1);
-
-                CHqlBoundExpr boundLower, boundUpper;
-                translator.buildExpr(subctx, rangeLower, boundLower);
-                translator.buildExpr(subctx, rangeUpper, boundUpper);
-                args.append(*getPointer(boundLower.expr));
-                args.append(*getPointer(boundUpper.expr));
-            }
-            else
-            {
-                OwnedHqlExpr address = getMonitorValueAddress(subctx, expandedSelector, normalized);
-                args.append(*LINK(address));
-                args.append(*LINK(address));
-            }
-            translator.callProcedure(subctx, func, args);
-        }
-    }
-    else
-    {
-        ForEachChild(idx2, values)
-        {
-            BuildCtx subctx(ctx);
-            IHqlExpression * cur = values->queryChild(idx2);
-            OwnedHqlExpr test = createBoolExpr(no_eq, LINK(lhs), LINK(cur));
-            OwnedHqlExpr promoted = getExplicitlyPromotedCompare(test);
-            OwnedHqlExpr compare, normalized;
-            extractCompareInformation(subctx, promoted, compare, normalized, expandedSelector);
-            if (compare)
-                translator.buildFilter(subctx, compare);
-
-            OwnedHqlExpr address = getMonitorValueAddress(subctx, expandedSelector, normalized);
-            HqlExprArray args;
-            args.append(*LINK(targetVar));
-            args.append(*LINK(address));
-            args.append(*LINK(address));
-
-            translator.callProcedure(subctx, func, args);
-        }
-    }
-}
-
-static IHqlExpression * createCompareRecast(node_operator op, IHqlExpression * value, IHqlExpression * recastValue)
-{
-    if (recastValue->queryValue())
-        return LINK(queryBoolExpr(op == no_ne));
-    return createValue(op, makeBoolType(), LINK(value), LINK(recastValue));
-}
-
-
-void MonitorExtractor::extractCompareInformation(BuildCtx & ctx, IHqlExpression * expr, SharedHqlExpr & compare, SharedHqlExpr & normalized, IHqlExpression * expandedSelector)
-{
-    extractCompareInformation(ctx, expr->queryChild(0), expr->queryChild(1), compare, normalized, expandedSelector);
-}
-
-
-void MonitorExtractor::extractCompareInformation(BuildCtx & ctx, IHqlExpression * lhs, IHqlExpression * value, SharedHqlExpr & compare, SharedHqlExpr & normalized, IHqlExpression * expandedSelector)
-{
-    LinkedHqlExpr compareValue = value->queryBody();
-    OwnedHqlExpr recastValue;
-    if ((lhs->getOperator() != no_select) || (lhs->queryType() != compareValue->queryType()))
-    {
-        OwnedHqlExpr temp  = castToFieldAndBack(lhs, compareValue);
-        if (temp != compareValue)
-        {
-            //Force into a temporary variable since it will be used more than once, and reapply the field casting/
-            compareValue.setown(translator.buildSimplifyExpr(ctx, compareValue));
-            //cast to promoted type because sometimes evaluating can convert string to string<n>
-            Owned<ITypeInfo> promotedType = getPromotedECLType(lhs->queryType(), compareValue->queryType());
-            compareValue.setown(ensureExprType(compareValue, promotedType));
-            recastValue.setown(castToFieldAndBack(lhs, compareValue));
-        }
-    }
-
-    normalized.setown(invertTransforms(lhs, compareValue));
-    normalized.setown(foldHqlExpression(normalized));
-
-    if (recastValue && recastValue != compareValue)
-        compare.setown(createCompareRecast(no_eq, compareValue, recastValue));
-}
-
-void MonitorExtractor::createStringSet(BuildCtx & ctx, const char * target, unsigned size, IHqlExpression * selector)
-{
-    assertex(selector->getOperator() == no_select);
-    if (createValueSets)
-    {
-        StringBuffer type;
-        translator.buildRtlFieldType(type, selector->queryChild(1), queryRecord(tableExpr));
-        ctx.addQuotedF("%s.setown(createValueSet(%s));", target, type.str());
-    }
-    else
-    {
-        if (onlyHozedCompares)
-            ctx.addQuotedF("%s.setown(createRtlStringSet(%u));", target, size);
-        else
-        {
-            ITypeInfo * type = selector->queryType();
-            bool isBigEndian = !type->isInteger() || !isLittleEndian(type);
-            ctx.addQuotedF("%s.setown(createRtlStringSetEx(%u,%d,%d));", target, size, isBigEndian, type->isSigned());
-        }
-    }
-}
-
-void MonitorExtractor::buildKeySegmentCompareExpr(BuildMonitorState & buildState, KeySelectorInfo & selectorInfo, BuildCtx & ctx, const char * targetSet, IHqlExpression & thisKey)
-{
-    OwnedHqlExpr targetVar = createVariable(targetSet, makeVoidType());
-    createStringSet(ctx, targetSet, selectorInfo.size, selectorInfo.expandedSelector);
-
-    if (!exprReferencesDataset(&thisKey, tableExpr))
-    {
-        BuildCtx subctx(ctx);
-        translator.buildFilter(subctx, &thisKey);
-        callAddAll(subctx, targetVar);
-        return;
-    }
-
-    OwnedHqlExpr compare;
-    OwnedHqlExpr normalized;
-    BuildCtx subctx(ctx);
-    extractCompareInformation(subctx, &thisKey, compare, normalized, selectorInfo.expandedSelector);
-    OwnedHqlExpr address = getMonitorValueAddress(subctx, selectorInfo.expandedSelector, normalized);
-
-    HqlExprArray args;
-    args.append(*LINK(targetVar));
-
-    node_operator op = thisKey.getOperator();
-    switch (op)
-    {
-    case no_eq:
-        if (compare)
-            translator.buildFilter(subctx, compare);
-        args.append(*LINK(address));
-        args.append(*LINK(address));
-        translator.callProcedure(subctx, addRangeFunc, args);
-        break;
-    case no_ne:
-        subctx.addQuoted(StringBuffer().appendf("%s->addAll();", targetSet));
-        if (compare)
-            translator.buildFilter(subctx, compare);
-        args.append(*LINK(address));
-        args.append(*LINK(address));
-        translator.callProcedure(subctx, killRangeFunc, args);
-        break;
-    case no_le:
-        args.append(*createValue(no_nullptr, makeVoidType()));
-        args.append(*LINK(address));
-        translator.callProcedure(subctx, addRangeFunc, args);
-        break;
-    case no_lt:
-        // e) no_lt.  If isExact add < value else add <= value
-        if (compare)
-        {
-            OwnedHqlExpr invCompare = getInverse(compare);
-            IHqlStmt * cond = translator.buildFilterViaExpr(subctx, invCompare);
-            //common this up...
-            args.append(*createValue(no_nullptr, makeVoidType()));
-            args.append(*LINK(address));
-            translator.callProcedure(subctx, addRangeFunc, args);
-            subctx.selectElse(cond);
-            args.append(*LINK(targetVar));
-        }
-        subctx.addQuoted(StringBuffer().appendf("%s->addAll();", targetSet));
-        args.append(*LINK(address));
-        args.append(*createValue(no_nullptr, makeVoidType()));
-        translator.callProcedure(subctx, killRangeFunc, args);
-        break;
-    case no_ge:
-        // d) no_ge.  If isExact add >= value else add > value
-        if (compare)
-        {
-            OwnedHqlExpr invCompare = getInverse(compare);
-            IHqlStmt * cond = translator.buildFilterViaExpr(subctx, invCompare);
-            //common this up...
-            subctx.addQuoted(StringBuffer().appendf("%s->addAll();", targetSet));
-            args.append(*createValue(no_nullptr, makeVoidType()));
-            args.append(*LINK(address));
-            translator.callProcedure(subctx, killRangeFunc, args);
-            subctx.selectElse(cond);
-            args.append(*LINK(targetVar));
-        }
-        args.append(*LINK(address));
-        args.append(*createValue(no_nullptr, makeVoidType()));
-        translator.callProcedure(subctx, addRangeFunc, args);
-        break;
-    case no_gt:
-        subctx.addQuoted(StringBuffer().appendf("%s->addAll();", targetSet));
-        args.append(*createValue(no_nullptr, makeVoidType()));
-        args.append(*LINK(address));
-        translator.callProcedure(subctx, killRangeFunc, args);
-        break;
-    case no_between:
-    case no_notbetween:
-        {
-            //NB: This should only be generated for substring queries.  User betweens are converted
-            //to two separate comparisons to cope with range issues.
-            args.append(*LINK(address));
-
-            CHqlBoundExpr rhs2;
-            OwnedHqlExpr adjustedUpper = invertTransforms(thisKey.queryChild(0), thisKey.queryChild(2));
-            OwnedHqlExpr foldedUpper = foldHqlExpression(adjustedUpper);
-            OwnedHqlExpr hozedValue = getHozedKeyValue(foldedUpper);
-            IIdAtom * name = hozedValue->queryId();
-            if ((name != createRangeHighId) && (name != createQStrRangeHighId))
-                hozedValue.setown(ensureExprType(hozedValue, selectorInfo.expandedSelector->queryType()));
-            translator.buildExpr(subctx, hozedValue, rhs2);
-            translator.ensureHasAddress(subctx, rhs2);
-            args.append(*getPointer(rhs2.expr));
-            if (op == no_between)
-                translator.callProcedure(subctx, addRangeFunc, args);
-            else
-            {
-                subctx.addQuoted(StringBuffer().appendf("%s->addAll();", targetSet));
-                translator.callProcedure(subctx, killRangeFunc, args);
-            }
-            break;
-        }
-    default:
-        throwUnexpectedOp(op);
-    }
-}
-
-
-IHqlExpression * MonitorExtractor::unwindConjunction(HqlExprArray & matches, IHqlExpression * expr)
-{
-    node_operator op = expr->getOperator();
-    expr->unwindList(matches, op);
-    OwnedHqlExpr invariant;
-    ForEachItemInRev(i, matches)
-    {
-        IHqlExpression & cur = matches.item(i);
-        if (isIndexInvariant(&cur, false))
-        {
-            invariant.setown(extendConditionOwn(op, LINK(&cur), invariant.getClear()));
-            matches.remove(i);
-        }
-    }
-    return invariant.getClear();
-}
-
-
-//Note this function may change the incoming ctx if filterKind is not NoMonitorFilter
-void MonitorExtractor::buildKeySegmentExpr(BuildMonitorState & buildState, KeySelectorInfo & selectorInfo, BuildCtx & ctx, const char * requiredSet, IHqlExpression & thisKey, MonitorFilterKind filterKind)
-{
-    const char * targetSet = requiredSet;
-    StringBuffer s;
-    unsigned curSize = selectorInfo.size;
-    node_operator op = thisKey.getOperator();
-
-    BuildCtx subctx(ctx);
-    BuildCtx * appendCtx = &ctx;
-    StringBuffer createMonitorText;
-
-    switch (op)
-    {
-    case no_in: 
-    case no_notin:
-        {
-            if (!targetSet)
-                targetSet = buildState.getSetName(createValueSets);
-            buildKeySegmentInExpr(buildState, selectorInfo, ctx, targetSet, thisKey, filterKind);
-            break;
-        }
-    case no_if:
-        {
-            MonitorFilterKind childFilter = targetSet ? NoMonitorFilter : filterKind;
-            IHqlStmt * ifStmt = translator.buildFilterViaExpr(subctx, thisKey.queryChild(0));
-            buildKeySegmentExpr(buildState, selectorInfo, subctx, targetSet, *thisKey.queryChild(1), childFilter);
-            subctx.selectElse(ifStmt);
-            buildKeySegmentExpr(buildState, selectorInfo, subctx, targetSet, *thisKey.queryChild(2), childFilter);
-            break;
-        }
-    case no_and:
-        {
-            HqlExprArray matches;
-            OwnedHqlExpr invariant = unwindConjunction(matches, &thisKey);
-            unsigned numMatches = matches.ordinality();
-
-            if (!targetSet && numMatches > 1)
-                targetSet = buildState.getSetName(createValueSets);
-
-            IHqlStmt * ifStmt = NULL;
-            if (invariant)
-            {
-                ifStmt = translator.buildFilterViaExpr(subctx, invariant);
-                if (filterKind == MonitorFilterSkipEmpty)
-                    ctx.set(subctx);
-            }
-            
-            buildKeySegmentExpr(buildState, selectorInfo, subctx, targetSet, matches.item(0), NoMonitorFilter);
-            for (unsigned i=1; i< numMatches; i++)
-            {
-                IHqlExpression & cur = matches.item(i);
-                const char * curTarget = buildState.getSetName(createValueSets);
-                BuildCtx childctx(subctx);
-                buildKeySegmentExpr(buildState, selectorInfo, childctx, curTarget, cur, MonitorFilterSkipAll);
-                if (createValueSets)
-                    childctx.addQuotedF("%s->intersectSet(%s);", targetSet, curTarget);
-                else
-                    childctx.addQuotedF("%s.setown(rtlIntersectSet(%s,%s));", targetSet, targetSet, curTarget);
-                buildState.popSetName();
-            }
-
-            if (invariant && (filterKind != MonitorFilterSkipEmpty))
-            {
-                subctx.selectElse(ifStmt);
-                if (targetSet)
-                    createStringSet(subctx, targetSet, curSize, selectorInfo.selector);
-                else
-                    buildEmptyKeySegment(buildState, subctx, selectorInfo);
-            }
-            break;
-        }
-    case no_or:
-        {
-            HqlExprArray matches;
-            OwnedHqlExpr invariant = unwindConjunction(matches, &thisKey);
-            unsigned numMatches = matches.ordinality();
-
-            if (invariant)
-            {
-                if (filterKind == MonitorFilterSkipAll)
-                {
-                    OwnedHqlExpr test = getInverse(invariant);
-                    translator.buildFilter(subctx, test);
-                    ctx.set(subctx);
-                }
-                else
-                {
-                    IHqlStmt * ifStmt = translator.buildFilterViaExpr(subctx, invariant);
-                    if (targetSet)
-                    {
-                        createStringSet(subctx, targetSet, curSize, selectorInfo.selector);
-                        OwnedHqlExpr targetVar = createVariable(targetSet, makeVoidType());
-                        callAddAll(subctx, targetVar);
-                    }
-                    subctx.selectElse(ifStmt);
-                }
-            }
-            
-            appendCtx = &subctx;
-            if (!targetSet && numMatches > 1)
-                targetSet = buildState.getSetName(createValueSets);
-
-            buildKeySegmentExpr(buildState, selectorInfo, subctx, targetSet, matches.item(0), NoMonitorFilter);
-            for (unsigned i=1; i < numMatches; i++)
-            {
-                IHqlExpression & cur = matches.item(i);
-                const char * curTarget = buildState.getSetName(createValueSets);
-                BuildCtx childctx(subctx);
-                buildKeySegmentExpr(buildState, selectorInfo, childctx, curTarget, cur, MonitorFilterSkipEmpty);
-                if (createValueSets)
-                    childctx.addQuotedF("%s->unionSet(%s);", targetSet, curTarget);
-                else
-                    childctx.addQuotedF("%s.setown(rtlUnionSet(%s, %s));", targetSet, targetSet, curTarget);
-                buildState.popSetName();
-            }
-            break;
-        }
-    case no_eq:
-        {
-            if (!targetSet)
-            {
-                if (buildSingleKeyMonitor(createMonitorText, selectorInfo, subctx, thisKey))
-                    break;
-                targetSet = buildState.getSetName(createValueSets);
-            }
-            buildKeySegmentCompareExpr(buildState, selectorInfo, ctx, targetSet, thisKey);
-            break;
-        }
-    default:
-        {
-            if (!targetSet)
-                targetSet = buildState.getSetName(createValueSets);
-            buildKeySegmentCompareExpr(buildState, selectorInfo, ctx, targetSet, thisKey);
-            break;
-        }
-    }
-
-    if (targetSet && !requiredSet)
-    {
-        if (createValueSets)
-            createMonitorText.appendf("createFieldFilter(%u, %s.getClear())", selectorInfo.fieldIdx, targetSet);
-        else
-            createMonitorText.appendf("createKeySegmentMonitor(%s, %s.getClear(), %u, %u, %u)",
-                                      boolToText(selectorInfo.keyedKind != KeyedYes), targetSet, selectorInfo.fieldIdx, selectorInfo.offset, selectorInfo.size);
-
-
-        buildState.popSetName();
-    }
-
-    if (createMonitorText.length())
-    {
-        if (createValueSets)
-            appendCtx->addQuotedF("%s->append(%s, %s);", buildState.listName, selectorInfo.getFFOptions(), createMonitorText.str());
-        else
-            appendCtx->addQuotedF("%s->append(%s);", buildState.listName, createMonitorText.str());
-    }
-}
-
-
-IHqlExpression * MonitorExtractor::getMonitorValueAddress(BuildCtx & ctx, IHqlExpression * selector, IHqlExpression * _value)
-{
-    LinkedHqlExpr value = _value;
-    CHqlBoundExpr bound;
-    ITypeInfo * type = selector->queryType();
-    bool castViaRow = isUnknownSize(type);
-    if (castViaRow)
-    {
-        IHqlExpression * field = selector->queryChild(1);
-        OwnedHqlExpr record = createRecord(field);
-        OwnedHqlExpr self = createSelector(no_self, record, nullptr);
-        OwnedHqlExpr assign = createValue(no_assign, makeVoidType(), createNewSelectExpr(LINK(self), LINK(field)), LINK(value));
-        OwnedHqlExpr transform = createValue(no_transform, makeTransformType(record->getType()), LINK(assign));
-        OwnedHqlExpr row = createRow(no_createrow, LINK(transform));
-        translator.buildAnyExpr(ctx, row, bound);
-    }
-    else
-    {
-        if (!createValueSets)
-        {
-            //Need to ensure old segmonitors for varstrings are filled with \0s
-            switch (type->getTypeCode())
-            {
-            case type_varstring: case type_varunicode:
-                {
-                    assertex(type->getSize() != UNKNOWN_LENGTH);
-                    CHqlBoundTarget tempTarget;
-                    translator.createTempFor(ctx, type, tempTarget, typemod_none, FormatNatural);
-                    //clear the variable.
-                    HqlExprArray args;
-                    args.append(*getPointer(tempTarget.expr));
-                    args.append(*getZero());
-                    args.append(*getSizetConstant(type->getSize()));
-                    OwnedHqlExpr call = translator.bindTranslatedFunctionCall(memsetId, args);
-                    ctx.addExpr(call);
-                    //then assign over the top
-                    translator.buildExprAssign(ctx, tempTarget, value);
-                    bound.setFromTarget(tempTarget);
-                    break;
-                }
-            }
-        }
-        if (!bound.expr)
-        {
-            translator.buildExpr(ctx, value, bound);
-            translator.ensureHasAddress(ctx, bound);
-        }
-    }
-    return getPointer(bound.expr);
-}
-
-bool MonitorExtractor::buildSingleKeyMonitor(StringBuffer & createMonitorText, KeySelectorInfo & selectorInfo, BuildCtx & ctx, IHqlExpression & thisKey)
-{
-    BuildCtx subctx(ctx);
-    OwnedHqlExpr compare, normalized;
-
-    StringBuffer funcName;
-    extractCompareInformation(subctx, &thisKey, compare, normalized, selectorInfo.expandedSelector);
-    if (compare)
-        return false;
-
-    if (createValueSets)
-    {
-        StringBuffer type;
-        translator.buildRtlFieldType(type, selectorInfo.selector->queryChild(1), tableExpr->queryRecord());
-
-        //MORE: Need to ensure it is exactly the correct format - e.g. variable length strings are length prefixed
-        OwnedHqlExpr address = getMonitorValueAddress(subctx, selectorInfo.expandedSelector, normalized);
-        StringBuffer addrText;
-        translator.generateExprCpp(addrText, address);
-
-        createMonitorText.appendf("createFieldFilter(%u, %s, %s)", selectorInfo.fieldIdx, type.str(), addrText.str());
-    }
-    else
-    {
-        ITypeInfo * type = selectorInfo.expandedSelector->queryType();
-        type_t tc = type->getTypeCode();
-        if ((tc == type_int) || (tc == type_swapint))
-        {
-            if (isLittleEndian(type))
-            {
-                if (type->isSigned())
-                    funcName.append("createSingleLittleSignedKeySegmentMonitor");
-                else if (type->getSize() != 1)
-                    funcName.append("createSingleLittleKeySegmentMonitor");
-            }
-            else
-            {
-                if (type->isSigned())
-                    funcName.append("createSingleBigSignedKeySegmentMonitor");
-                else
-                    funcName.append("createSingleKeySegmentMonitor");
-            }
-        }
-
-        if (!funcName.length())
-            funcName.append("createSingleKeySegmentMonitor");
-
-        OwnedHqlExpr address = getMonitorValueAddress(subctx, selectorInfo.expandedSelector, normalized);
-        StringBuffer addrText;
-        translator.generateExprCpp(addrText, address);
-
-        createMonitorText.append(funcName)
-                         .appendf("(%s, %u, %u, %u, %s)",
-                                  boolToText(selectorInfo.keyedKind != KeyedYes), selectorInfo.fieldIdx, selectorInfo.offset, selectorInfo.size, addrText.str());
-    }
-    return true;
-}
-
-KeyedKind getKeyedKind(HqlCppTranslator & translator, KeyConditionArray & matches)
-{
-    KeyedKind keyedKind = KeyedNo;
-    ForEachItemIn(i, matches)
-    {
-        KeyCondition & cur = matches.item(i);
-        if (cur.keyedKind != keyedKind)
-        {
-            if (keyedKind == KeyedNo)
-                keyedKind = cur.keyedKind;
-            else
-                translator.throwError1(HQLERR_InconsistentKeyedOpt, str(cur.selector->queryChild(1)->queryName()));
-        }
-    }
-    return keyedKind;
-}
-
-void MonitorExtractor::buildEmptyKeySegment(BuildMonitorState & buildState, BuildCtx & ctx, KeySelectorInfo & selectorInfo)
-{
-    StringBuffer s;
-    if (createValueSets)
-    {
-        StringBuffer type;
-        translator.buildRtlFieldType(type, selectorInfo.selector->queryChild(1), queryRecord(selectorInfo.selector->queryChild(0)));
-        ctx.addQuoted(s.appendf("%s->append(%s, createEmptyFieldFilter(%u, %s));", buildState.listName, selectorInfo.getFFOptions(), selectorInfo.fieldIdx, type.str()));
-    }
-    else
-        ctx.addQuoted(s.appendf("%s->append(createEmptyKeySegmentMonitor(%s, %u, %u, %u));", buildState.listName, boolToText(selectorInfo.keyedKind != KeyedYes), selectorInfo.fieldIdx, selectorInfo.offset, selectorInfo.size));
-}
-
-
-void MonitorExtractor::buildKeySegment(BuildMonitorState & buildState, BuildCtx & ctx, unsigned whichField, unsigned curSize)
-{
-    IHqlExpression * selector = &keyableSelects.item(whichField);
-    IHqlExpression * expandedSelector = &expandedSelects.item(whichField);
-    IHqlExpression * field = selector->queryChild(1);
-    KeyConditionArray matches;
-    bool isImplicit = true;
-    bool prevWildWasKeyed = buildState.wildWasKeyed;
-    buildState.wildWasKeyed = false;
-    ForEachItemIn(cond, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(cond);
-        if (cur.selector == selector)
-        {
-            cur.generated = true;
-            if (cur.isWild)
-            {
-                isImplicit = false;
-                if (cur.wasKeyed)
-                    buildState.wildWasKeyed = true;
-                else if (buildState.implicitWildField && !ignoreUnkeyed)
-                {
-                    StringBuffer s, keyname;
-                    translator.throwError3(HQLERR_WildFollowsGap, getExprECL(field, s).str(), str(buildState.implicitWildField->queryChild(1)->queryName()), queryKeyName(keyname));
-                }
-            }
-            else
-            {
-                matches.append(OLINK(cur));
-                if (buildState.implicitWildField && !ignoreUnkeyed)
-                {
-                    StringBuffer s,keyname;
-                    if (cur.isKeyed())
-                        translator.throwError3(HQLERR_KeyedFollowsGap, getExprECL(field, s).str(), str(buildState.implicitWildField->queryChild(1)->queryName()), queryKeyName(keyname));
-                    else if (!buildState.doneImplicitWarning)
-                    {
-                        translator.WARNING3(CategoryEfficiency, HQLWRN_KeyedFollowsGap, getExprECL(field, s).str(), str(buildState.implicitWildField->queryChild(1)->queryName()), queryKeyName(keyname));
-                        buildState.doneImplicitWarning = true;
-                    }
-                }
-            }
-        }
-    }
-    if (buildState.wildWasKeyed && (matches.ordinality() == 0))
-    {
-        StringBuffer keyname;
-        translator.WARNING2(CategoryFolding, HQLWRN_FoldRemoveKeyed, str(field->queryName()), queryKeyName(keyname));
-    }
-
-    StringBuffer s;
-    KeyedKind keyedKind = getKeyedKind(translator, matches);
-    if (whichField >= firstOffsetField)
-        translator.throwError1(HQLERR_KeyedNotKeyed, getExprECL(field, s).str());
-    KeySelectorInfo selectorInfo(keyedKind, selector, expandedSelector, buildState.curFieldIdx, buildState.curOffset, curSize);
-
-    bool ignoreKeyedExtend = false;
-    if ((keyedKind == KeyedExtend) && buildState.wildPending() && !ignoreUnkeyed)
-    {
-        if (keyedKind == KeyedExtend)
-        {
-            if (prevWildWasKeyed)
-                buildState.wildWasKeyed = true;
-            else
-            {
-                StringBuffer keyname;
-                translator.WARNING2(CategoryEfficiency, HQLERR_OptKeyedFollowsWild, getExprECL(field, s).str(), queryKeyName(keyname));
-            }
-        }
-        //previous condition folded so always true, so keyed,opt will always be a wildcard.
-
-        if (!allowDynamicFormatChange)
-            ignoreKeyedExtend = true;
-        isImplicit = false;
-    }
-
-    if (matches.ordinality() && !ignoreKeyedExtend)
-    {
-        if (buildState.wildPending() && !ignoreUnkeyed)
-            buildState.clearWild();
-
-        HqlExprArray args;
-        ForEachItemIn(i, matches)
-        {
-            KeyCondition & cur = matches.item(i);
-            args.append(*LINK(cur.expr));
-        }
-
-        OwnedHqlExpr fullExpr = createBalanced(no_and, queryBoolType(), args);
-        BuildCtx subctx(ctx);
-        buildKeySegmentExpr(buildState, selectorInfo, subctx, NULL, *fullExpr, ignoreUnkeyed ? MonitorFilterSkipAll : NoMonitorFilter);
-    }
-    else
-    {
-        if (isImplicit)
-        {
-            buildState.implicitWildField.set(selector);
-            buildState.doneImplicitWarning = false;
-        }
-
-        if (buildState.wildPending() && noMergeSelects.contains(*selector))
-            buildState.clearWild();
-
-        if (!buildState.wildPending())
-            buildState.wildOffset = buildState.curOffset;
-    }
-    buildState.curOffset += selectorInfo.size;
-    buildState.curFieldIdx++;
-}
-
-void MonitorExtractor::spotSegmentCSE(BuildCtx & ctx)
-{
-    //This could make things much better, but needs some thought
-    HqlExprArray conditions;
-    ForEachItemIn(cond, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(cond);
-        if (cur.expr)
-            conditions.append(*LINK(cur.expr));
-    }
-
-    HqlExprArray associated;
-    IHqlExpression * selector = tableExpr->queryNormalizedSelector();
-    translator.traceExpressions("before seg spot", conditions);
-    spotScalarCSE(conditions, associated, NULL, selector, translator.queryOptions().spotCseInIfDatasetConditions);
-    translator.traceExpressions("after seg spot", conditions);
-
-    unsigned curCond = 0;
-    ForEachItemIn(i, conditions)
-    {
-        IHqlExpression * cur = &conditions.item(i);
-        switch (cur->getOperator())
-        {
-        case no_alias:
-            translator.buildStmt(ctx, cur);
-            break;
-        case no_alias_scope:
-            translator.expandAliasScope(ctx, cur);
-            cur = cur->queryChild(0);
-            //fallthrough
-        default:
-            for (;;)
-            {   
-                if (!keyed.conditions.isItem(curCond))
-                    throwUnexpected();
-                KeyCondition & keyCond = keyed.conditions.item(curCond++);
-                if (keyCond.expr)
-                {
-                    keyCond.expr.set(cur);
-                    break;
-                }
-            }
-            break;
-        }
-    }
-    for (;;)
-    {   
-        if (!keyed.conditions.isItem(curCond))
-            break;
-        KeyCondition & keyCond = keyed.conditions.item(curCond++);
-        assertex(!keyCond.expr);
-    }
-}
-
-
-void MonitorExtractor::buildSegments(BuildCtx & ctx, const char * listName, bool _ignoreUnkeyed)
-{
-    translator.useInclude("rtlkey.hpp");
-    ignoreUnkeyed = _ignoreUnkeyed;
-
-    if (translator.queryOptions().spotCSE)
-        spotSegmentCSE(ctx);
-
-    BuildMonitorState buildState(ctx, listName);
-    ForEachItemIn(idx, keyableSelects)
-    {
-        IHqlExpression * selector = &keyableSelects.item(idx);
-        IHqlExpression * expandedSelector = &expandedSelects.item(idx);
-        IHqlExpression * field = selector->queryChild(1);
-        unsigned curSize = expandedSelector->queryType()->getSize();
-        assertex(createValueSets || curSize != UNKNOWN_LENGTH);
-
-        //MORE: Should also allow nested record structures, and allow keying on first elements.
-        //      and field->queryType()->getSize() doesn't work for alien datatypes etc.
-        if(!field->hasAttribute(virtualAtom))
-            buildKeySegment(buildState, ctx, idx, curSize);
-    }
-
-    //check that all keyed entries have been matched
-    ForEachItemIn(cond, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(cond);
-        if (!cur.generated)
-            translator.throwError(HQLERR_OnlyKeyFixedField);
-    }
-}
-
-
-bool MonitorExtractor::isKeySelect(IHqlExpression * select)
-{
-    return (keyableSelects.find(*select) != NotFound);
-}
-
-bool MonitorExtractor::isEqualityFilter(IHqlExpression * search)
-{
-    bool matched = false;
-    ForEachItemIn(cond, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(cond);
-        if (cur.selector == search)
-        {
-            if (!cur.isWild)
-            {
-                if (matched)
-                    return false;
-                matched = true;
-                IHqlExpression * matchExpr = cur.expr;
-                if (matchExpr->getOperator() != no_eq)
-                    return false;
-            }
-        }
-    }
-    return matched;
-}
-
-bool MonitorExtractor::isEqualityFilterBefore(IHqlExpression * select)
-{
-    ForEachItemIn(i, keyableSelects)
-    {
-        IHqlExpression & cur = keyableSelects.item(i);
-        if (select == &cur)
-            return true;
-        if (!isEqualityFilter(&cur))
-            return false;
-    }
-    throwUnexpected();
-}
-
-bool MonitorExtractor::isPrevSelectKeyed(IHqlExpression * select)
-{
-    unsigned match = keyableSelects.find(*select);
-    assertex(match != NotFound);
-    if (match == 0)
-        return true;
-    IHqlExpression * prev = &keyableSelects.item(match-1);
-    ForEachItemIn(i, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(i);
-        if (cur.selector == prev)
-        {
-            if (!cur.isWild && cur.isKeyed())
-                return true;
-            if (cur.wasKeyed)
-                return true;
-        }
-    }
-    return false;
-}
-
-
-bool MonitorExtractor::okToKey(IHqlExpression * select, KeyedKind keyedKind)
-{
-    if (keyedKind == KeyedYes)
-        return true;
-
-    ForEachItemIn(i, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(i);
-        if (cur.selector == select && cur.isWild)
-            return false;
-    }
-
-    return true;
-}
-
-bool MonitorExtractor::isIndexInvariant(IHqlExpression * expr, bool includeRoot)
-{
-    if (containsAssertKeyed(expr))
-        return false;
-
-    HqlExprCopyArray scopeUsed;
-    expr->gatherTablesUsed(NULL, &scopeUsed);
-
-    IHqlExpression * search = tableExpr->queryNormalizedSelector();
-    ForEachItemIn(i, scopeUsed)
-    {
-        IHqlExpression * cur = &scopeUsed.item(i);
-        for (;;)
-        {
-            if (cur == search)
-                return false;
-
-            if (includeRoot && (queryRoot(cur) == search))
-                return false;
-
-            IHqlExpression * parent = queryNextMultiLevelDataset(cur, true);
-            if (!parent)
-                break;
-            cur = parent;
-        }
-    }
-    return true;
-}
-
-
-
-
-IHqlExpression * MonitorExtractor::castToFieldAndBack(IHqlExpression * left, IHqlExpression * right)
-{
-    node_operator op = left->getOperator();
-    switch (op)
-    {
-    case no_cast:
-    case no_implicitcast:
-        {
-            IHqlExpression * uncast = left->queryChild(0);
-            ITypeInfo * castType = right->queryType();
-            ITypeInfo * uncastType = uncast->queryType();
-
-            OwnedHqlExpr castRight = ensureExprType(right, uncastType);
-            OwnedHqlExpr base = castToFieldAndBack(uncast, castRight);
-            //If this cast doesn't lose any information and child didn't change then don't bother
-            //casting back and forwards.
-            if ((base == castRight) && !castLosesInformation(uncastType, castType))
-                return LINK(right);
-            return ensureExprType(base, castType);
-        }
-
-    case no_select:
-        {
-            ITypeInfo * leftType = left->queryType();
-            ITypeInfo * rightType = right->queryType();
-            if (leftType == rightType || !castLosesInformation(leftType, rightType))
-                return LINK(right);
-            return ensureExprType(right, leftType);
-        }
-    case no_substring:
-    case no_add:
-    case no_sub:
-        return castToFieldAndBack(left->queryChild(0), right);
-    default:
-        UNIMPLEMENTED;
-    }
-}
-
-
-
-IHqlExpression * MonitorExtractor::invertTransforms(IHqlExpression * left, IHqlExpression * right)
-{
-    node_operator op = left->getOperator();
-    switch (op)
-    {
-    case no_cast:
-    case no_implicitcast:
-        {
-            assertex(right->queryType()->getTypeCode() != type_set);
-
-            IHqlExpression * uncast = left->queryChild(0);
-            ITypeInfo * uncastType = uncast->queryType();
-
-
-            OwnedHqlExpr castRight = ensureExprType(right, uncastType);
-            return invertTransforms(uncast, castRight);
-        }
-
-    case no_select:
-        {
-            assertex(isKeySelect(left));
-            ITypeInfo * leftType = left->queryType();
-            ITypeInfo * rightType = right->queryType();
-            if (leftType == rightType || !castLosesInformation(leftType, rightType))
-                return LINK(right);
-            return ensureExprType(right, leftType);
-        }
-
-    case no_add:
-    case no_sub:
-        {
-            assertex(right->getOperator() != no_list);
-
-            OwnedHqlExpr adjusted = createValue(op == no_sub ? no_add : no_sub, right->getType(), LINK(right), LINK(left->queryChild(1)));
-            return invertTransforms(left->queryChild(0), adjusted);
-        }
-     case no_substring:
-         {
-             assertex(right->getOperator() != no_list);
-
-             return invertTransforms(left->queryChild(0), right);
-         }
-    default:
-        UNIMPLEMENTED;
-    }
-}
-
-
-
-IHqlExpression * MonitorExtractor::queryKeyableSelector(IHqlExpression * expr)
-{
-    switch (expr->getOperator())
-    {
-    case no_cast:
-    case no_implicitcast:
-    case no_add:
-    case no_sub:
-        return queryKeyableSelector(expr->queryChild(0));
-    case no_select:
-        if (isKeySelect(expr))
-            return expr;
-        return NULL;
-    }
-    return NULL;
-}
-
-IHqlExpression * MonitorExtractor::isKeyableFilter(IHqlExpression * left, IHqlExpression * right, bool & duplicate, node_operator compareOp, KeyFailureInfo & reason, KeyedKind keyedKind)
-{
-    node_operator op = left->getOperator();
-    switch (op)
-    {
-    case no_cast:
-    case no_implicitcast:
-        {
-            IHqlExpression * uncast = left->queryChild(0);
-            ITypeInfo * castType = left->queryType();
-            ITypeInfo * uncastType = uncast->queryType();
-            //(ty)x = y.            E.g., (int1)string2field = int1value
-            //if more than one value of x[uncastType] corresponds to a single value in y[castType] then we can't sensibly create
-            //the key segment monitor.  Because we will get false negatives.  If it is an inverse then duplicate (see below)
-            bool canRemoveCast = true;
-            if (castLosesInformation(castType, uncastType))
-            {
-                if ((compareOp != no_ne) && (compareOp != no_notin))
-                    canRemoveCast = false;
-                duplicate = true;
-            }
-
-            //if more than one value of y corresponds to a single value of x then need to duplicate the test condition.
-            //or pretest whether (ty)(tx)y == y.
-            //if (castLosesInformation(uncastType, castType))
-            //Now taken care of when the segment monitors are created
-
-            //If the comparison is non equality and the cast changes the collation sequence then you can't remove it.
-            switch (compareOp)
-            {
-            case no_eq:
-            case no_ne:
-            case no_in:
-            case no_notin:
-                break;
-            default:
-                if (!preservesOrder(castType, uncastType))
-                    canRemoveCast = false;
-                break;
-            }
-
-            Linked<ITypeInfo> newType = uncastType;
-            if (right->queryType()->getTypeCode() == type_set)
-                newType.setown(makeSetType(newType.getLink()));
-            OwnedHqlExpr castRight = ensureExprType(right, newType);
-            IHqlExpression * ret = isKeyableFilter(uncast, castRight, duplicate, compareOp, reason, keyedKind);
-            if (canRemoveCast || !ret)
-                return ret;
-            reason.set(KFRcast, ret->queryChild(1));
-            return NULL;
-        }
-
-    case no_select:
-        if (isKeySelect(left) && okToKey(left, keyedKind))
-        {
-            if (isIndexInvariant(right, false))
-                return left;
-            reason.set(KFRtoocomplex, left);
-        }
-        else
-            reason.set(KFRnokey);
-        return NULL;
-
-    case no_substring:
-        {
-            IHqlExpression * range = left->queryChild(1);
-            if (range->getOperator() == no_rangeto)
-            {
-                IValue *end = range->queryChild(0)->queryValue();
-                if (!end)
-                    break;
-                return isKeyableFilter(left->queryChild(0), right, duplicate, compareOp, reason, keyedKind);
-            }
-            else if (range->getOperator() == no_range)
-            {
-                IValue *start = range->queryChild(0)->queryValue();
-                IValue *end = range->queryChild(1)->queryValue();
-                if (!start || !end || start->getIntValue() != 1)
-                    break;
-                return isKeyableFilter(left->queryChild(0), right, duplicate, compareOp, reason, keyedKind);
-            }
-            reason.set(KFRtoocomplex, right);
-            return NULL;
-        }
-    case no_add:
-    case no_sub:
-        if (isIndexInvariant(left->queryChild(1), false))
-            return isKeyableFilter(left->queryChild(0), right, duplicate, compareOp, reason, keyedKind);
-        reason.set(KFRtoocomplex, left);
-        return NULL;
-    }
-    reason.set(KFRnokey);
-    return NULL;
-}
-
-
-
-static IHqlExpression * getCompareValue(ITypeInfo * fieldType, unsigned subStringLen, IValue * value, int whichBoundary)
-{
-    type_t ftc = fieldType->getTypeCode();
-    unsigned fieldLen = fieldType->getStringLen();
-    unsigned lenValue = value->queryType()->getStringLen();
-    const void * rawValue = value->queryValue();
-    size32_t resultLen;
-    rtlDataAttr result;
-
-    if (whichBoundary < 0)
-    {
-        switch (ftc)
-        {
-        case type_qstring:
-            rtlCreateQStrRangeLow(resultLen, result.refstr(), fieldLen, subStringLen, lenValue, static_cast<const char *>(rawValue));
-            break;
-        case type_string:
-            rtlCreateStrRangeLow(resultLen, result.refstr(), fieldLen, subStringLen, lenValue, static_cast<const char *>(rawValue));
-            break;
-        case type_data:
-            rtlCreateDataRangeLow(resultLen, result.refdata(), fieldLen, subStringLen, lenValue, rawValue);
-            break;
-        case type_unicode:
-            rtlCreateUnicodeRangeLow(resultLen, result.refustr(), fieldLen, subStringLen, lenValue, static_cast<const UChar *>(rawValue));
-            break;
-        default:
-            //should this generate a warning/error instead?
-            rtlCreateRange(resultLen, result.refstr(), fieldLen, subStringLen, fieldType->getSize(), static_cast<const char *>(rawValue), 0, 0);
-            break;
-        }
-    }
-    else
-    {
-        switch (ftc)
-        {
-        case type_qstring:
-            rtlCreateQStrRangeHigh(resultLen, result.refstr(), fieldLen, subStringLen, lenValue, static_cast<const char *>(rawValue));
-            break;
-        case type_string:
-            rtlCreateStrRangeHigh(resultLen, result.refstr(), fieldLen, subStringLen, lenValue, static_cast<const char *>(rawValue));
-            break;
-        case type_data:
-            rtlCreateDataRangeHigh(resultLen, result.refdata(), fieldLen, subStringLen, lenValue, rawValue);
-            break;
-        case type_unicode:
-            rtlCreateUnicodeRangeHigh(resultLen, result.refustr(), fieldLen, subStringLen, lenValue, static_cast<const UChar *>(rawValue));
-            break;
-        default:
-            rtlCreateRange(resultLen, result.refstr(), fieldLen, subStringLen, fieldType->getSize(), static_cast<const char *>(rawValue), 255, 0);
-            break;
-        }
-    }
-    assertex(resultLen == fieldLen);
-    return createConstant(createValueFromMem(LINK(fieldType), result.getdata()));
-}
-
-
-IHqlExpression * MonitorExtractor::getRangeLimit(ITypeInfo * fieldType, IHqlExpression * lengthExpr, IHqlExpression * value, int whichBoundary)
-{
-    type_t ftc = fieldType->getTypeCode();
-    unsigned fieldLength = fieldType->getStringLen();
-    IValue * constValue = value->queryValue();
-
-    if (constValue && lengthExpr->queryValue())
-    {
-        unsigned subStringLen = (unsigned)lengthExpr->queryValue()->getIntValue();
-        if ((int)subStringLen < 0) subStringLen = 0;
-        if (subStringLen > fieldLength)
-            translator.throwError1(HQLERR_SubstringOutOfRange, subStringLen);
-
-        return getCompareValue(fieldType, subStringLen, constValue, whichBoundary);
-    }
-
-    IIdAtom * func;
-    if (whichBoundary < 0)
-    {
-        switch (ftc)
-        {
-        case type_qstring:
-            func = createQStrRangeLowId;
-            break;
-        case type_string:
-            func = createStrRangeLowId;
-            break;
-        case type_data:
-            func = createDataRangeLowId;
-            break;
-        case type_unicode:
-            func = createUnicodeRangeLowId;
-            break;
-        default:
-            func = createRangeLowId;
-            break;
-        }
-    }
-    else
-    {
-        switch (ftc)
-        {
-        case type_qstring:
-            func = createQStrRangeHighId;
-            break;
-        case type_string:
-            func = createStrRangeHighId;
-            break;
-        case type_data:
-            func = createDataRangeHighId;
-            break;
-        case type_unicode:
-            func = createUnicodeRangeHighId;
-            break;
-        default:
-            func = createRangeHighId;
-            break;
-        }
-    }
-
-    HqlExprArray args;
-    args.append(*getSizetConstant(fieldLength));
-    args.append(*LINK(lengthExpr));
-    args.append(*LINK(value));
-
-    //Note: I can't change the return type of the function - because if fixed length then wrong call is made, and variable length is worse code.
-    OwnedHqlExpr call = translator.bindFunctionCall(func, args);
-    return createValue(no_typetransfer, LINK(fieldType), LINK(call));
-}
-
-
-IHqlExpression * MonitorExtractor::createRangeCompare(IHqlExpression * selector, IHqlExpression * value, IHqlExpression * lengthExpr, bool compareEqual)
-{
-    OwnedHqlExpr foldedValue = foldHqlExpression(value);
-    ITypeInfo * fieldType = selector->queryType();
-    IHqlExpression * lowExpr = getRangeLimit(fieldType, lengthExpr, foldedValue, -1);
-    IHqlExpression * highExpr = getRangeLimit(fieldType, lengthExpr, foldedValue, +1);
-
-    //Could convert to two separate tests, but code is worse, and boundary conditions aren't going to happen.
-    return createValue(compareEqual ? no_between : no_notbetween, makeBoolType(), LINK(selector), lowExpr, highExpr);
-}
-
-
-static IHqlExpression * removeCastTrim(IHqlExpression * expr)
-{
-    for (;;)
-    {
-        if ((expr->getOperator() == no_trim) && !expr->queryChild(1))
-            expr = expr->queryChild(0);
-        else if (isLengthPreservingCast(expr))
-            expr = expr->queryChild(0);
-        else
-            return expr;
-
-        expr = queryNonAliased(expr);
-    }
-}
-
-static IHqlExpression * queryLengthFromRange(IHqlExpression * range)
-{
-    switch (range->getOperator())
-    {
-    case no_rangefrom:
-    case no_rangecommon:
-        return NULL;
-    case no_rangeto:
-        return range->queryChild(0);
-    case no_range:
-        if (getIntValue(range->queryChild(0), 0) == 1)
-            return range->queryChild(1);
-        return NULL;
-    default:
-        if (getIntValue(range, 0) == 1)
-            return range;
-        return NULL;
-    }
-}
-
-static void extendRangeCheck(SharedHqlExpr & globalGuard, SharedHqlExpr & localCond, IHqlExpression * selector, IHqlExpression * lengthExpr, bool compareEqual)
-{
-#if 0
-    //This might be a good idea, but probably doesn't make a great deal of difference at runtime
-    //Optimize the case where you check zero length to use a wild carded range instead
-    //x[1..len=0] = y from x in range 0000000..FFFFFF to  len==0 || x in range....
-    if (compareEqual)
-    {
-        if (!lengthExpr->queryValue())
-        {
-            OwnedHqlExpr testLength = createBoolExpr(no_eq, LINK(lengthExpr), ensureExprType(queryZero(), lengthExpr->queryType()));
-            localCond.setown(createBoolExpr(no_or, testLength.getClear(), LINK(localCond)));
-        }
-    }
-#endif
-    //For a range check x[1..m] = y we generate
-    //x in range(sizeof(x), m, y, 0) and range(maxlength(x), m, y, 255)
-    //need to guard with condition len(trim(y)) <= m
-
-    //If x[1..length(trim(y))] == y then don't add a condition
-    IHqlExpression * cur = queryNonAliased(lengthExpr);
-    IHqlExpression * compare = removeCastTrim(queryNonAliased(selector));
-    if (cur->getOperator() == no_charlen)
-    {
-        cur = queryNonAliased(cur->queryChild(0));
-        if (cur->getOperator() == no_trim)
-        {
-            cur = queryNonAliased(cur->queryChild(0));
-            while (isLengthPreservingCast(cur))
-                cur = queryNonAliased(cur->queryChild(0));
-
-            IHqlExpression * compare = selector;
-            for (;;)
-            {
-                compare = removeCastTrim(queryNonAliased(compare));
-                if (cur->queryBody() == compare->queryBody())
-                    return;
-                //Casts between strings that reduce/increase the number of characters don't matter as long as
-                //they eventually match the search string
-                if (!isCast(compare) || !isStringType(compare->queryType()))
-                    break;
-                compare = compare->queryChild(0);
-                if (!isStringType(compare->queryType()))
-                    break;
-            }
-        }
-    }
-
-    // if x[1..n] = z[1..n] then no need for a condition
-    if (compare->getOperator() == no_substring)
-    {
-        IHqlExpression * range = queryLengthFromRange(compare->queryChild(1));
-        if (range == lengthExpr)
-            return;
-    }
-
-    // if x[1..n] = (string<m>y) where m<n then no need for condition
-    unsigned selectorLength = selector->queryType()->getStringLen();
-    if (selectorLength <= getIntValue(lengthExpr, 0))
-        return;
-
-    //otherwise, if x[1..y] == z then add check length(trim(z)) <= y
-    OwnedHqlExpr trim = createTrimExpr(selector, NULL);
-    OwnedHqlExpr len = createValue(no_charlen, LINK(unsignedType), LINK(trim));
-    ITypeInfo * lengthType = lengthExpr->queryType();
-    Owned<ITypeInfo> compareType = getPromotedECLCompareType(unsignedType, lengthType);
-    node_operator compOp = compareEqual ? no_le : no_gt;
-    OwnedHqlExpr positiveLen = createValue(no_maxlist, lengthExpr->getType(), createValue(no_list, makeSetType(LINK(lengthType)), LINK(lengthExpr), createConstant(lengthType->castFrom(false, I64C(0)))));
-    OwnedHqlExpr test = createValue(no_le, makeBoolType(), ensureExprType(len, compareType), ensureExprType(positiveLen, compareType));
-    test.setown(foldHqlExpression(test));
-    if (compareEqual)
-        extendConditionOwn(globalGuard, no_and, test.getClear());
-    else
-        extendConditionOwn(localCond, no_or, getInverse(test));
-}
-
-
-bool MonitorExtractor::matchSubstringFilter(KeyConditionInfo & matches, node_operator op, IHqlExpression * left, IHqlExpression * right, KeyedKind keyedKind, bool & duplicate)
-{
-    LinkedHqlExpr value = right;
-    duplicate = false;
-    OwnedHqlExpr guard;
-    ITypeInfo * guardCastType = NULL;
-
-    if ((left->getOperator() == no_cast) || (left->getOperator() == no_implicitcast))
-    {
-        //code is extracted and simplified from isKeyableFilter() above - should be commoned up.
-        IHqlExpression * uncast = left->queryChild(0);
-        ITypeInfo * castType = left->queryType();
-        ITypeInfo * uncastType = uncast->queryType();
-
-        //(ty)x = y.
-        //if more than one value of x[uncastType] corresponds to a single value in y[castType] then we can't sensibly create
-        //the key segment monitor.  Because we will get false negatives.  If it is an inverse then duplicate (see below)
-        bool canRemoveCast = true;
-        if (castLosesInformation(castType, uncastType))
-            canRemoveCast = false;
-
-        //if more than one value of y corresponds to a single value of x then need to duplicate the test condition.
-        if (!preservesOrder(castType, uncastType))
-            canRemoveCast = false;
-
-        if (!canRemoveCast)
-        {
-//          reason.set(KFRcast, ret->queryChild(1));
-            return false;
-        }
-
-        if ((op != no_in) && (op != no_notin))
-        {
-            value.setown(ensureExprType(right, uncastType));
-
-            //If a simple equality test then create a global guard to check that we aren't matching a false positive
-            if (castLosesInformation(uncastType, castType))
-                guard.setown(createBoolExpr(no_eq, ensureExprType(value, castType), LINK(right)));
-        }
-        else
-        {
-            //if an IN then add guards to each comparison - generated later...
-            if (castLosesInformation(uncastType, castType))
-                guardCastType = uncastType;
-            else
-            {
-                Owned<ITypeInfo> targetType = makeSetType(LINK(uncastType));
-                value.setown(ensureExprType(right, targetType));
-            }
-        }
-
-        left = uncast;
-    }
-
-    if (left->getOperator() != no_substring)
-        return false;
-
-    if ((op == no_in) || (op == no_notin))
-    {
-        value.setown(normalizeListCasts(value));
-        if (value->getOperator() != no_list)
-            return false;
-    }
-
-    IHqlExpression * selector = left->queryChild(0);
-    if (!isKeySelect(selector) || !okToKey(selector, keyedKind))
-        return false;
-    if (!isIndexInvariant(right, false))
-        return false;
-    ITypeInfo * fieldType = selector->queryType();
-    unsigned fieldLength = fieldType->getStringLen();
-    if (fieldLength == UNKNOWN_LENGTH)
-        return false;
-
-    OwnedHqlExpr range = foldHqlExpression(left->queryChild(1));
-    IHqlExpression * lengthExpr = queryLengthFromRange(range);
-    if (!lengthExpr)
-        return false;
-
-    OwnedHqlExpr newTest;
-    if ((op == no_eq) || (op == no_ne))
-    {
-        newTest.setown(createRangeCompare(selector, value, lengthExpr, op == no_eq));
-        extendRangeCheck(guard, newTest, right, lengthExpr, op == no_eq);
-    }
-    else //no_in, no_notin
-    {
-        HqlExprArray compares;
-        ForEachChild(i, value)
-        {
-            IHqlExpression * cur = value->queryChild(i);
-            LinkedHqlExpr castValue = cur;
-            OwnedHqlExpr valueGuard;
-
-            if (guardCastType)
-            {
-                castValue.setown(ensureExprType(castValue, guardCastType));
-                valueGuard.setown(createBoolExpr(no_eq, ensureExprType(castValue, cur->queryType()), LINK(cur)));
-                extendRangeCheck(valueGuard, valueGuard, cur, lengthExpr, (op == no_in));
-            }
-            OwnedHqlExpr cond = createRangeCompare(selector, castValue, lengthExpr, (op == no_in));
-            if (valueGuard)
-                cond.setown(createValue(no_and, makeBoolType(), valueGuard.getClear(), cond.getClear()));
-            compares.append(*cond.getClear());
-        }
-
-        node_operator combineOp = (op == no_in) ? no_or : no_and;
-        newTest.setown(createBalanced(combineOp, queryBoolType(), compares));
-    }
-    
-    matches.appendCondition(*new KeyCondition(selector, newTest, keyedKind));
-    if (guard)
-        matches.appendPreFilter(guard);
-    return true;
-}
-
-
-bool MonitorExtractor::extractSimpleCompareFilter(KeyConditionInfo & matches, IHqlExpression * expr, KeyedKind keyedKind) 
-{
-    OwnedHqlExpr promoted = getExplicitlyPromotedCompare(expr);
-    IHqlExpression * l = promoted->queryChild(0);
-    IHqlExpression * r = promoted->queryChild(1);
-    bool duplicate = false;
-    KeyFailureInfo reasonl, reasonr;
-    node_operator op = expr->getOperator();
-    IHqlExpression * matchedSelector = isKeyableFilter(l, r, duplicate, op, reasonl, keyedKind);
-    Owned<KeyCondition> result;
-    if (matchedSelector)
-    {
-        node_operator newOp = getModifiedOp(op, duplicate);
-        if (newOp != no_none)
-        {
-            OwnedHqlExpr newFilter = createValue(newOp, expr->getType(), LINK(l), LINK(r));
-            result.setown(new KeyCondition(matchedSelector, newFilter, keyedKind));
-        }
-    }
-    else
-    {
-        duplicate = false;
-        matchedSelector = isKeyableFilter(r, l, duplicate, op, reasonr, keyedKind);
-        if (matchedSelector)
-        {
-            node_operator newOp = getModifiedOp(getReverseOp(op), duplicate);
-            if (newOp != no_none)
-            {
-                OwnedHqlExpr newFilter = createValue(newOp, expr->getType(), LINK(r), LINK(l));
-                result.setown(new KeyCondition(matchedSelector, newFilter, keyedKind));
-            }
-        }
-    }
-
-    bool extracted = (result != NULL);
-    if (extracted)
-    {
-        matches.appendCondition(*result.getClear());
-    }
-    else
-    {
-        failReason.merge(reasonl);
-        failReason.merge(reasonr);
-    }
-    if (duplicate || !extracted)
-        matches.appendPostFilter(expr);
-    return extracted;
-}
-
-
-bool MonitorExtractor::extractOrFilter(KeyConditionInfo & matches, IHqlExpression * expr, KeyedKind keyedKind)
-{
-    HqlExprArray conds;
-    expr->unwindList(conds, no_or);
-
-    bool validOrFilter = true;
-    HqlExprAttr invariant;
-    CIArrayOf<KeyConditionInfo> branches;
-    ForEachItemIn(idx, conds)
-    {
-        IHqlExpression & cur = conds.item(idx);
-        if (isIndexInvariant(&cur, false))
-            extendOrCondition(invariant, &cur);
-        else
-        {
-            KeyConditionInfo & branch = * new KeyConditionInfo;
-            branches.append(branch);
-            //Can't generate an OR with a pure post-filter
-            if (!extractFilters(branch, &cur, keyedKind))
-                validOrFilter = false;
-        }
-    }
-
-    //check all the conditions that are ORd together don't contain references to multiple fields.
-    KeyCondition * firstBranch = NULL;
-    bool multipleBranches = branches.ordinality() > 1;
-    bool multipleSelectors = false;
-    bool multipleConditions = false;
-    bool hasPostFilter = false;
-    ForEachItemIn(i1, branches)
-    {
-        KeyConditionInfo & branch = branches.item(i1);
-        if (branch.postFilter)
-            hasPostFilter = true;
-        ForEachItemIn(i2, branch.conditions)
-        {
-            KeyCondition & cur = branch.conditions.item(i2);
-            if (!firstBranch)
-                firstBranch = &cur;
-            else
-            {
-                multipleConditions = true;
-                if (i1 == 0)
-                {
-                    if (firstBranch->selector != cur.selector)
-                        multipleSelectors = true;
-                }
-                else
-                {
-                    if ((firstBranch->selector != cur.selector) || multipleSelectors)
-                        validOrFilter = false;
-                }
-            }
-        }
-    }
-
-    if (multipleBranches && hasPostFilter)
-        validOrFilter = false;
-
-    if (validOrFilter && firstBranch)
-    {
-        bool optimizeSingleBranch = true;
-        if (multipleSelectors || hasPostFilter || (optimizeSingleBranch && !multipleConditions))
-        {
-            //Invariant ored with a conjunction
-            //X or (A and B) -> (X or A) AND (X or B)
-            assertex(branches.ordinality() == 1);
-            KeyConditionInfo & branch = branches.item(0);
-            OwnedHqlExpr preFilter = branch.preFilter ? extendCondition(no_or, invariant, branch.preFilter) : NULL;
-            OwnedHqlExpr postFilter = branch.postFilter ? extendCondition(no_or, invariant, branch.postFilter) : NULL;
-
-            matches.appendPreFilter(preFilter);
-            matches.appendPostFilter(postFilter);
-            ForEachItemIn(i2, branch.conditions)
-            {
-                KeyCondition & cur = branch.conditions.item(i2);
-                OwnedHqlExpr filter = extendCondition(no_or, invariant, cur.expr);
-                matches.conditions.append(*new KeyCondition(cur.selector, filter, keyedKind));
-            }
-        }
-        else
-        {
-            LinkedHqlExpr combinedCondition = invariant;
-
-            ForEachItemIn(i1, branches)
-            {
-                KeyConditionInfo & branch = branches.item(i1);
-                OwnedHqlExpr conjunction = branch.createConjunction();
-                extendOrCondition(combinedCondition, conjunction);
-            }
-            
-            matches.conditions.append(*new KeyCondition(firstBranch->selector, combinedCondition, keyedKind));
-        }
-        return true;
-    }
-    else
-    {
-        matches.appendPostFilter(expr);
-        KeyFailureInfo reason;
-        reason.set(KFRor);
-        failReason.merge(reason);
-        return false;
-    }
-}
-
-
-bool MonitorExtractor::extractIfFilter(KeyConditionInfo & matches, IHqlExpression * expr, KeyedKind keyedKind)
-{
-    //MORE: This could generate better code, but I don't think it is worth the effort at the moment.
-    //Really, I should analyse left and right.  Iterate each selector referenced.  If there are no post conditions then
-    //generate IF(a, X, Y) compound expression, otherwise generate the default below.
-    IHqlExpression * cond = expr->queryChild(0);
-    if ((keyedKind != KeyedNo) && isIndexInvariant(cond, false))
-    {
-        //Convert IF(a, X, Y) to... IF (a, X, true) AND IF (a, true, Y) to... (NOT a OR X) AND (a OR Y)
-        OwnedHqlExpr inverseCond = getInverse(cond);
-        OwnedHqlExpr trueValue = createBoolExpr(no_or, LINK(inverseCond), LINK(expr->queryChild(1)));
-        OwnedHqlExpr falseValue = createBoolExpr(no_or, LINK(cond), LINK(expr->queryChild(2)));
-        OwnedHqlExpr combined = createBoolExpr(no_and, LINK(trueValue), LINK(falseValue));
-        return extractFilters(matches, combined, keyedKind);
-    }
-    matches.appendPostFilter(expr);
-    return false;
-}
-
-
-static HqlTransformerInfo selectSpotterInfo("SelectSpotter");
-MonitorExtractor::SelectSpotter::SelectSpotter(const HqlExprArray & _selects) : NewHqlTransformer(selectSpotterInfo), selects(_selects) 
-{ 
-    hasSelects = false; 
-}
-
-void MonitorExtractor::SelectSpotter::analyseExpr(IHqlExpression * expr)
-{
-    if (hasSelects || alreadyVisited(expr))
-        return;
-    if (selects.find(*expr) != NotFound)
-    {
-        hasSelects = true;
-        return;
-    }
-    NewHqlTransformer::analyseExpr(expr);
-}
-
-bool MonitorExtractor::containsTableSelects(IHqlExpression * expr)
-{
-    HqlExprCopyArray inScope;
-    expr->gatherTablesUsed(NULL, &inScope);
-
-    //Check that cursors for all inScope tables are already bound in the start context
-    return inScope.find(*tableExpr->queryNormalizedSelector()) != NotFound;
-}
-
-
-void MonitorExtractor::extractFilters(IHqlExpression * expr, SharedHqlExpr & extraFilter)
-{
-    HqlExprArray conds;
-    expr->unwindList(conds, no_and);
-    extractFilters(conds, extraFilter);
-}
-
-void MonitorExtractor::extractFilters(HqlExprArray & exprs, SharedHqlExpr & extraFilter)
-{
-    OwnedHqlExpr savedFilter = keyed.postFilter.getClear();
-    ForEachItemIn(i1, exprs)
-    {
-        IHqlExpression & cur = exprs.item(i1);
-        switch (cur.getOperator())
-        {
-        case no_assertkeyed:
-        case no_assertwild:
-            extractFilters(keyed, &cur, KeyedNo);
-            break;
-        }
-    }
-
-    keyedExplicitly = (keyed.conditions.ordinality() != 0);
-    cleanlyKeyedExplicitly = keyedExplicitly && !keyed.postFilter;
-    ForEachItemIn(i2, exprs)
-    {
-        IHqlExpression & cur = exprs.item(i2);
-        switch (cur.getOperator())
-        {
-        case no_assertkeyed:
-        case no_assertwild:
-            break;
-        default:
-            if (!keyedExplicitly)
-                extractFilters(keyed, &cur, KeyedNo);
-            else if (!cur.isAttribute() && isIndexInvariant(&cur, true))
-                keyed.appendPreFilter(&cur);
-            else
-                keyed.appendPostFilter(&cur);
-            break;
-        }
-    }
-
-    extraFilter.set(keyed.postFilter);
-    keyed.postFilter.setown(extendConditionOwn(no_and, savedFilter.getClear(), LINK(extraFilter)));
-}
-
-
-void MonitorExtractor::extractFiltersFromFilterDs(IHqlExpression * expr)
-{
-    HqlExprArray conds;
-    HqlExprAttr dummy;
-    unwindFilterConditions(conds, expr);
-    extractFilters(conds, dummy);
-}
-
-
-
-void MonitorExtractor::extractFoldedWildFields(IHqlExpression * expr)
-{
-    node_operator op = expr->getOperator();
-    switch (op)
-    {
-    case no_cast:
-    case no_implicitcast:
-    case no_add:
-    case no_sub:
-        //fields may have been transformed since folding...
-        extractFoldedWildFields(expr->queryChild(0));
-        break;
-    case no_select:
-        if (isKeySelect(expr))
-        {
-            KeyCondition * condition = new KeyCondition;
-            condition->selector.set(expr);
-            condition->isWild = true;
-            condition->wasKeyed = true;
-            keyed.conditions.append(*condition);
-        }
-        break;
-    }
-}
-
-bool MonitorExtractor::extractBoolFieldFilter(KeyConditionInfo & matches, IHqlExpression * selector, KeyedKind keyedKind, bool compareValue)
-{
-    if (selector->isBoolean())
-    {
-        if (isKeySelect(selector) && okToKey(selector, keyedKind))
-        {
-            OwnedHqlExpr newFilter = createValue(no_eq, makeBoolType(), LINK(selector), createConstant(compareValue));
-            matches.appendCondition(*new KeyCondition(selector, newFilter, keyedKind));
-            return true;
-        }
-    }
-    return false;
-}
-
-bool MonitorExtractor::extractFilters(KeyConditionInfo & matches, IHqlExpression * expr, KeyedKind keyedKind)
-{
-    if (!expr->isAttribute() && isIndexInvariant(expr, true))
-    {
-        extendAndCondition(matches.preFilter, expr);
-        return true;
-    }
-
-    IHqlExpression *l = expr->queryChild(0);
-    IHqlExpression *r = expr->queryChild(1);
-    node_operator op = expr->getOperator();
-    switch (op)
-    {
-    case no_and:
-        {
-            bool extracted = extractFilters(matches, l, keyedKind);
-            if (!extractFilters(matches, r, keyedKind)) extracted = false;
-            return extracted;
-        }
-    case no_or:
-        return extractOrFilter(matches, expr, keyedKind);
-    case no_attr:
-    case no_attr_expr:
-    case no_attr_link:
-        return true;
-    case no_not:
-        {
-            IHqlExpression * arg = expr->queryChild(0);
-            OwnedHqlExpr inverse = getInverse(arg);
-            if (inverse->queryBody() != expr->queryBody())
-                return extractFilters(matches, inverse, keyedKind);
-
-            if ((arg->getOperator() == no_select) && arg->isBoolean() &&
-                extractBoolFieldFilter(matches, arg, keyedKind, false))
-                return true;
-
-            matches.appendPostFilter(expr);
-            return false;
-        }
-    case no_between:
-    case no_notbetween:
-        {
-            //Convert this into two comparisons because that will handle weird boundary conditions much better.
-            OwnedHqlExpr normalized = expandBetween(expr);
-            return extractFilters(matches, normalized, keyedKind);
-        }
-    case no_eq:
-    case no_ne:
-        {
-            bool duplicate = false;
-            if (matchSubstringFilter(matches, op, l, r, keyedKind, duplicate) || matchSubstringFilter(matches, op, r, l, keyedKind, duplicate))
-            {
-                if (duplicate)
-                    matches.appendPostFilter(expr);
-                return true;
-            }
-            return extractSimpleCompareFilter(matches, expr, keyedKind);
-        }
-    case no_in:
-    case no_notin:
-        {
-            bool duplicate = false;
-            if (matchSubstringFilter(matches, op, l, r, keyedKind, duplicate))
-            {
-                if (duplicate)
-                    matches.appendPostFilter(expr);
-                return true;
-            }
-            return extractSimpleCompareFilter(matches, expr, keyedKind);
-        }
-    case no_gt:
-    case no_lt:
-    case no_ge:
-    case no_le:
-        return extractSimpleCompareFilter(matches, expr, keyedKind);
-    case no_assertkeyed:
-        {
-            KeyFailureInfo reason;
-            reason.merge(failReason);
-            failReason.clear();
-            bool extend = expr->hasAttribute(extendAtom);
-            if (!extractFilters(matches, l, extend ? KeyedExtend : KeyedYes))
-            {
-                if (!extend)
-                    failReason.reportError(translator, expr);
-            }
-
-            IHqlExpression * original = expr->queryAttribute(_selectors_Atom);
-            if (original)
-            {
-                ForEachChild(i, original)
-                    extractFoldedWildFields(original->queryChild(i));
-            }
-            failReason.merge(reason);
-            return true;
-        }
-    case no_assertwild:
-        {
-            if (l->getOperator() == no_all)
-            {
-                IHqlExpression * original = expr->queryAttribute(_selectors_Atom);
-                assertex(original);
-                ForEachChild(i, original)
-                    extractFoldedWildFields(original->queryChild(i));
-            }
-            else
-            {
-                IHqlExpression * selector = queryKeyableSelector(l);
-                if (!selector)
-                {
-                    StringBuffer keyname;
-                    translator.throwError1(HQLERR_WildNotReferenceIndex, queryKeyName(keyname));
-                }
-                KeyCondition * condition = new KeyCondition;
-                condition->selector.set(selector);
-                condition->isWild = true;
-                matches.appendCondition(*condition);
-            }
-            return true;
-        }
-    case no_if:
-        return extractIfFilter(matches, expr, keyedKind);
-    case no_select:
-        {
-            if (expr->isBoolean() && extractBoolFieldFilter(matches, expr, keyedKind, true))
-                return true;
-
-            matches.appendPostFilter(expr);
-            return false;
-        }
-    default:
-        // Add this condition to the catchall expr
-        matches.appendPostFilter(expr);
-        return false;
-    }
-}
-
-
-void MonitorExtractor::extractAllFilters(IHqlExpression * dataset)
-{
-    for (;;)
-    {
-        switch (dataset->getOperator())
-        {
-        case no_newkeyindex:
-            return;
-        case no_filter:
-            extractAllFilters(dataset->queryChild(0));
-            extractFiltersFromFilterDs(dataset);
-            return;
-        case no_compound_indexread:
-        case no_newusertable:
-        case no_hqlproject:
-        case no_distributed:
-        case no_preservemeta:
-        case no_unordered:
-        case no_sorted:
-        case no_stepped:
-        case no_grouped:
-        case no_alias_scope:
-        case no_dataset_alias:
-            break;
-        default:
-            UNIMPLEMENTED;
-        }
-        dataset = dataset->queryChild(0);
-    }
-}
-
-
-bool MonitorExtractor::isKeyed()                                            
-{ 
-    ForEachItemIn(i, keyed.conditions)
-    {
-        if (!keyed.conditions.item(i).isWild)
-            return true;
-    }
-    return false;
-}
-
-
-bool expandFilename(StringBuffer & s, IHqlExpression * expr)
-{
-    switch (expr->getOperator())
-    {
-    case no_constant:
-        expr->toString(s);
-        return true;
-    case no_getresult:
-        //more
-        break;
-    case no_concat:
-        {
-            bool hadString = expandFilename(s, expr->queryChild(0));
-            unsigned oldLength = s.length();
-            s.append("+");
-            if (expandFilename(s, expr->queryChild(1)) || hadString)
-                return true;
-            s.setLength(oldLength);
-            return false;
-        }
-    case no_alias:
-    case no_cast:
-    case no_implicitcast:
-        return expandFilename(s, expr->queryChild(0));
-    }
-    if (hasNamedSymbol(expr))
-    {
-        s.append(expr->queryName());
-        return true;
-    }
-    s.append("...");
-    return false;
-}
-
-const char * MonitorExtractor::queryKeyName(StringBuffer & s)
-{
-    IAtom * name = tableExpr->queryName();
-    if (name)
-        s.append(" \'").append(name).append("'");
-    else
-    {
-        IHqlExpression * filename = queryTableFilename(tableExpr);
-        if (filename)
-        {
-            if (!expandFilename(s.append(' '), filename))
-                s.clear();
-        }
-    }
-    return s.str();
-}
-
-
-static bool isNextField(IHqlExpression * record, IHqlExpression * prevExpr, IHqlExpression * nextExpr)
-{
-    if ((prevExpr->getOperator() != no_select) || (nextExpr->getOperator() != no_select))
-        return false;
-    IHqlExpression * prevSelect = prevExpr->queryChild(0);
-    if (prevSelect != nextExpr->queryChild(0))
-        return false;
-    if (!containsOnlyLeft(prevExpr))
-        return false;
-    if (prevSelect->getOperator() != no_left)
-        record = prevSelect->queryRecord();
-
-    IHqlExpression * prevField = prevExpr->queryChild(1);
-    IHqlExpression * nextField = nextExpr->queryChild(1);
-
-    //Slow, but probably doesn't matter...
-    ForEachChild(i, record)
-    {
-        IHqlExpression * cur = record->queryChild(i);
-        if (cur == prevField)
-            return (record->queryChild(i+1) == nextField);
-    }
-    return false;
-}
-
-
-IHqlExpression * MonitorExtractor::querySimpleJoinValue(IHqlExpression * selector)
-{
-    IHqlExpression * matched = NULL;
-    ForEachItemIn(cond, keyed.conditions)
-    {
-        KeyCondition & cur = keyed.conditions.item(cond);
-        if (cur.selector == selector)
-        {
-            if (!cur.isWild)
-            {
-                if (matched)
-                    return NULL;
-                IHqlExpression * matchExpr = cur.expr;
-                if (matchExpr->getOperator() != no_eq)
-                    return NULL;
-                if (matchExpr->queryChild(0) != selector)
-                    return NULL;
-                matched = matchExpr->queryChild(1);
-            }
-        }
-    }
-
-    return matched;
-}
-
 //---------------------------------------------------------------------------
 
 
@@ -6188,7 +3600,7 @@ class IndexReadBuilderBase : public SourceBuilder
     friend class MonitorRemovalTransformer;
 public:
     IndexReadBuilderBase(HqlCppTranslator & _translator, IHqlExpression *_tableExpr, IHqlExpression *_nameExpr)
-        : SourceBuilder(_translator, _tableExpr, _nameExpr), monitors(_tableExpr, _translator, -(int)numPayloadFields(_tableExpr), false)
+        : SourceBuilder(_translator, _tableExpr, _nameExpr), monitors(_tableExpr, _translator, -(int)numPayloadFields(_tableExpr), false, false)
     { 
     }
 
@@ -6204,7 +3616,7 @@ protected:
     IHqlExpression * removeMonitors(IHqlExpression * expr);
 
 protected:
-    MonitorExtractor monitors;
+    CppFilterExtractor monitors;
     SourceSteppingInfo steppingInfo;
 };
