+++ rtl/eclrtl/rtlfield.cpp
@@ -18,12 +18,15 @@
 #include "platform.h"
 #include <math.h>
 #include <stdio.h>
+#include <atomic>
 #include "jmisc.hpp"
 #include "jlib.hpp"
 #include "eclhelper.hpp"
 #include "eclrtl_imp.hpp"
 #include "rtlfield.hpp"
 #include "rtlds_imp.hpp"
+#include "rtlrecord.hpp"
+#include "rtlkey.hpp"
 #include "nbcd.hpp"
 
 static const char * queryXPath(const RtlFieldInfo * field)
@@ -251,6 +254,11 @@ const RtlFieldInfo * const * RtlTypeInfoBase::queryFields() const
 const RtlTypeInfo * RtlTypeInfoBase::queryChildType() const 
 {
     return NULL; 
+}
+
+const IFieldFilter * RtlTypeInfoBase::queryFilter() const
+{
+    return nullptr;
 }
 
 size32_t RtlTypeInfoBase::deserialize(ARowBuilder & builder, IRowDeserializerSource & in, size32_t offset) const
@@ -2849,8 +2857,10 @@ size32_t RtlRecordTypeInfo::deserialize(ARowBuilder & builder, IRowDeserializerS
 
 void RtlRecordTypeInfo::readAhead(IRowPrefetcherSource & in) const
 {
-    //Will not generally be called because it will have been expanded
-    return readAheadFields(fields, in);
+    //This could be called if the record contains ifblocks
+    in.noteStartChild();
+    readAheadFields(fields, in);
+    in.noteFinishChild();
 }
 
 
@@ -3802,12 +3812,75 @@ unsigned RtlIfBlockTypeInfo::hash(const byte * self, unsigned inhash) const
     return inhash;
 }
 
-bool RtlDynamicIfBlockTypeInfo::getCondition(const byte * selfrow) const
+bool RtlComplexIfBlockTypeInfo::getCondition(const RtlRow & selfrow) const
 {
-#ifdef _DEBUG
-    // Temporary code to help my testing, until proper implementation available
-    return selfrow[3] != 2;
-#endif
+    //Call the function in the derived class that evaluates the test condition
+    return getCondition(selfrow.queryRow());
+}
+
+
+static CriticalSection ifcs;
+RtlSerialIfBlockTypeInfo::~RtlSerialIfBlockTypeInfo()
+{
+    ::Release(filter);
+    delete parentRecord;
+}
+
+void RtlSerialIfBlockTypeInfo::doCleanup() const
+{
+    delete parentRecord;
+    parentRecord = nullptr;
+    ::Release(filter);
+    filter = nullptr;
+    RtlIfBlockTypeInfo::doCleanup();
+}
+
+
+bool RtlSerialIfBlockTypeInfo::getCondition(const byte * selfrow) const
+{
+    const IFieldFilter * filter = resolveCondition();
+    RtlDynRow row(*parentRecord, nullptr);
+    //Can only expand offset for fields up to and including this if block - otherwise it will call getCondition() again...
+    row.setRow(selfrow, numPrevFields);
+    return filter->matches(row);
+}
+
+bool RtlSerialIfBlockTypeInfo::getCondition(const RtlRow & selfrow) const
+{
+    return resolveCondition()->matches(selfrow);
+}
+
+const IFieldFilter * RtlSerialIfBlockTypeInfo::queryFilter() const
+{
+    return resolveCondition();
+}
+
+const IFieldFilter * RtlSerialIfBlockTypeInfo::resolveCondition() const
+{
+    //The following cast is a hack to avoid <atomic> being included from the header
+    //if parentRecord is not initialised then may need to create the filter condition
+    typedef std::atomic<RtlRecord *> AtomicRtlRecord;
+    AtomicRtlRecord & atomicParentRecord = *reinterpret_cast<AtomicRtlRecord *>(&parentRecord);
+    RtlRecord * parent = atomicParentRecord.load(std::memory_order_relaxed);
+    if (!parent)
+    {
+        CriticalBlock block(ifcs);
+        parent = atomicParentRecord.load(std::memory_order_relaxed);
+        if (!parent)
+        {
+            if (!filter)
+                filter = createCondition();
+            parent = new RtlRecord(*rowType, true);
+            atomicParentRecord.store(parent, std::memory_order_relaxed);
+            numPrevFields = parentRecord->queryIfBlockLimit(this);
+        }
+    }
+    return filter;
+}
+
+IFieldFilter * RtlDynamicIfBlockTypeInfo::createCondition() const
+{
+    //The filter should be initialised on deserialization
     UNIMPLEMENTED;
 }
