+++ testing/regress/ecl/key/nestedif.xml
@@ -4,3 +4,9 @@
  <Row><id>3</id><numnested>3</numnested><nestedrecord><Row><hasforename>true</hasforename><hassurname>false</hassurname><forename>Jim</forename></Row><Row><hasforename>false</hasforename><hassurname>true</hassurname><surname>Jones</surname></Row><Row><hasforename>false</hasforename><hassurname>false</hassurname></Row></nestedrecord></Row>
  <Row><id>99</id><numnested>0</numnested><nestedrecord></nestedrecord></Row>
 </Dataset>
+<Dataset name='Result 2'>
+ <Row><id>1</id><first><hasforename>true</hasforename><hassurname>true</hassurname><forename>Gavin</forename><surname>Hawthorn</surname></first><second><hasforename>false</hasforename><hassurname>false</hassurname></second></Row>
+ <Row><id>2</id><first><hasforename>false</hasforename><hassurname>false</hassurname></first><second><hasforename>true</hasforename><hassurname>false</hassurname><forename>X</forename></second></Row>
+ <Row><id>3</id><first><hasforename>true</hasforename><hassurname>false</hassurname><forename>Jim</forename></first><second><hasforename>false</hasforename><hassurname>true</hassurname><surname>Jones</surname></second></Row>
+ <Row><id>99</id><first><hasforename>false</hasforename><hassurname>false</hassurname></first><second><hasforename>false</hasforename><hassurname>false</hassurname></second></Row>
+</Dataset>
