+++ testing/regress/ecl/key/serializetypes.xml
@@ -722,3 +722,231 @@
 <Dataset name='Result 6'>
  <Row><Result_6>true</Result_6></Row>
 </Dataset>
+<Dataset name='Result 7'>
+ <Row><Result_7>{
+ &quot;ty1&quot;: {
+  &quot;fieldType&quot;: 257,
+  &quot;length&quot;: 8
+ },
+ &quot;ty2&quot;: {
+  &quot;fieldType&quot;: 0,
+  &quot;length&quot;: 1
+ },
+ &quot;ty3&quot;: {
+  &quot;fieldType&quot;: 1028,
+  &quot;length&quot;: 0
+ },
+ &quot;ty4&quot;: {
+  &quot;fieldType&quot;: 1067,
+  &quot;length&quot;: 0,
+  &quot;filterField&quot;: 0,
+  &quot;filterType&quot;: &quot;ty2&quot;,
+  &quot;filter&quot;: &quot;*&quot;,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;forename&quot;,
+    &quot;type&quot;: &quot;ty3&quot;,
+    &quot;flags&quot;: 1028
+   }
+  ]
+ },
+ &quot;ty5&quot;: {
+  &quot;fieldType&quot;: 1037,
+  &quot;length&quot;: 9,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;hasforename&quot;,
+    &quot;type&quot;: &quot;ty2&quot;
+   },
+   {
+    &quot;name&quot;: &quot;surnamecount&quot;,
+    &quot;type&quot;: &quot;ty1&quot;,
+    &quot;flags&quot;: 257
+   },
+   {
+    &quot;type&quot;: &quot;ty4&quot;,
+    &quot;flags&quot;: 1067
+   },
+   {
+    &quot;type&quot;: &quot;ty4&quot;,
+    &quot;flags&quot;: 1067
+   }
+  ]
+ },
+ &quot;ty6&quot;: {
+  &quot;fieldType&quot;: 532,
+  &quot;length&quot;: 0,
+  &quot;child&quot;: &quot;ty5&quot;
+ },
+ &quot;ty7&quot;: {
+  &quot;fieldType&quot;: 1067,
+  &quot;length&quot;: 0,
+  &quot;filterField&quot;: 0,
+  &quot;filterType&quot;: &quot;ty1&quot;,
+  &quot;filter&quot;: &quot;*&quot;,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;extra&quot;,
+    &quot;type&quot;: &quot;ty3&quot;,
+    &quot;flags&quot;: 1028
+   }
+  ]
+ },
+ &quot;fieldType&quot;: 1037,
+ &quot;length&quot;: 38,
+ &quot;fields&quot;: [
+  {
+   &quot;name&quot;: &quot;id&quot;,
+   &quot;type&quot;: &quot;ty1&quot;,
+   &quot;flags&quot;: 257
+  },
+  {
+   &quot;name&quot;: &quot;first&quot;,
+   &quot;type&quot;: &quot;ty5&quot;,
+   &quot;flags&quot;: 1037
+  },
+  {
+   &quot;name&quot;: &quot;second&quot;,
+   &quot;type&quot;: &quot;ty5&quot;,
+   &quot;flags&quot;: 1037
+  },
+  {
+   &quot;name&quot;: &quot;nestedrecord&quot;,
+   &quot;type&quot;: &quot;ty6&quot;,
+   &quot;xpath&quot;: &quot;nestedrecord\u0001Row&quot;,
+   &quot;flags&quot;: 532
+  },
+  {
+   &quot;type&quot;: &quot;ty7&quot;,
+   &quot;flags&quot;: 1067
+  }
+ ]
+}</Result_7></Row>
+</Dataset>
+<Dataset name='Result 8'>
+ <Row><Result_8>{
+ &quot;ty1&quot;: {
+  &quot;fieldType&quot;: 257,
+  &quot;length&quot;: 8
+ },
+ &quot;ty2&quot;: {
+  &quot;fieldType&quot;: 0,
+  &quot;length&quot;: 1
+ },
+ &quot;ty3&quot;: {
+  &quot;fieldType&quot;: 1028,
+  &quot;length&quot;: 0
+ },
+ &quot;ty4&quot;: {
+  &quot;fieldType&quot;: 1067,
+  &quot;length&quot;: 0,
+  &quot;filterField&quot;: 0,
+  &quot;filterType&quot;: &quot;ty2&quot;,
+  &quot;filter&quot;: &quot;=[&apos;1&apos;]&quot;,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;forename&quot;,
+    &quot;type&quot;: &quot;ty3&quot;,
+    &quot;flags&quot;: 1028
+   }
+  ]
+ },
+ &quot;ty5&quot;: {
+  &quot;fieldType&quot;: 1067,
+  &quot;length&quot;: 0,
+  &quot;filterField&quot;: 1,
+  &quot;filterType&quot;: &quot;ty1&quot;,
+  &quot;filter&quot;: &quot;=(,0),(0,)&quot;,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;surname&quot;,
+    &quot;type&quot;: &quot;ty3&quot;,
+    &quot;flags&quot;: 1028
+   }
+  ]
+ },
+ &quot;ty6&quot;: {
+  &quot;fieldType&quot;: 1037,
+  &quot;length&quot;: 9,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;hasforename&quot;,
+    &quot;type&quot;: &quot;ty2&quot;
+   },
+   {
+    &quot;name&quot;: &quot;surnamecount&quot;,
+    &quot;type&quot;: &quot;ty1&quot;,
+    &quot;flags&quot;: 257
+   },
+   {
+    &quot;type&quot;: &quot;ty4&quot;,
+    &quot;flags&quot;: 1067
+   },
+   {
+    &quot;type&quot;: &quot;ty5&quot;,
+    &quot;flags&quot;: 1067
+   }
+  ]
+ },
+ &quot;ty7&quot;: {
+  &quot;fieldType&quot;: 532,
+  &quot;length&quot;: 0,
+  &quot;child&quot;: &quot;ty6&quot;
+ },
+ &quot;ty8&quot;: {
+  &quot;fieldType&quot;: 1067,
+  &quot;length&quot;: 0,
+  &quot;filterField&quot;: 0,
+  &quot;filterType&quot;: &quot;ty1&quot;,
+  &quot;filter&quot;: &quot;=[1],[3],[5],[7],[12]&quot;,
+  &quot;fields&quot;: [
+   {
+    &quot;name&quot;: &quot;extra&quot;,
+    &quot;type&quot;: &quot;ty3&quot;,
+    &quot;flags&quot;: 1028
+   }
+  ]
+ },
+ &quot;fieldType&quot;: 1037,
+ &quot;length&quot;: 38,
+ &quot;fields&quot;: [
+  {
+   &quot;name&quot;: &quot;id&quot;,
+   &quot;type&quot;: &quot;ty1&quot;,
+   &quot;flags&quot;: 257
+  },
+  {
+   &quot;name&quot;: &quot;first&quot;,
+   &quot;type&quot;: &quot;ty6&quot;,
+   &quot;flags&quot;: 1037
+  },
+  {
+   &quot;name&quot;: &quot;second&quot;,
+   &quot;type&quot;: &quot;ty6&quot;,
+   &quot;flags&quot;: 1037
+  },
+  {
+   &quot;name&quot;: &quot;nestedrecord&quot;,
+   &quot;type&quot;: &quot;ty7&quot;,
+   &quot;xpath&quot;: &quot;nestedrecord\u0001Row&quot;,
+   &quot;flags&quot;: 532
+  },
+  {
+   &quot;type&quot;: &quot;ty8&quot;,
+   &quot;flags&quot;: 1067
+  }
+ ]
+}</Result_8></Row>
+</Dataset>
+<Dataset name='Result 9'>
+ <Row><Result_9>false</Result_9></Row>
+</Dataset>
+<Dataset name='Result 10'>
+ <Row><Result_10>505F136D36389A4F76C30000000101000008000000000104040000002B0440000000012A01666F72656E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000032B0400001402100000042B0440000000002A0165787472610002040400000D04400026056964000001010000666972737400040D0400007365636F6E6400040D0400006E65737465647265636F72640005140280006E65737465647265636F726401526F770000062B040000</Result_10></Row>
+</Dataset>
+<Dataset name='Result 11'>
+ <Row><Result_11>5026EF98B914CB032C4E0100000101000008000000000104040000002B0440000000013D020601030101666F72656E616D650002040400002B0440000001003D048C01000000000000000004000000000000000091017375726E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000042B0400001402100000052B0440000000003D0A060100000000000000030100000000000000060300000000000000030300000000000000060500000000000000030500000000000000060700000000000000030700000000000000060C00000000000000030C000000000000000165787472610002040400000D04400026056964000001010000666972737400050D0400007365636F6E6400050D0400006E65737465647265636F72640006140280006E65737465647265636F726401526F770000072B040000</Result_11></Row>
+</Dataset>
+<Dataset name='Result 12'>
+ <Row><Result_12>false</Result_12></Row>
+</Dataset>
