+++ testing/regress/ecl/nestedif.ecl
@@ -52,3 +52,30 @@ saved4 := TABLE(saved3, { saved3, unsigned forceTransform := 0; }) : independent
 saved5 := TABLE(saved4, { saved4 } - [forceTransform]);
 
 output(saved5);
+
+
+//test that the same record type embedded twice in the same record resolves the child fields correctly
+
+baseRecord := RECORD
+    UNSIGNED id;
+    nestedRecord first;
+    nestedRecord second;
+END;
+
+baseTable := dataset([
+        {1, {true, true, 'Gavin', 'Hawthorn'}, { false, false } },
+        {2, {false, false}, { true, false, 'X' } },
+        {3, {true, false, 'Jim'}, { false, true, 'Jones' } },
+        {99, {false, false}, {false, false} }], baseRecord);
+
+bsaved1 := baseTable : independent(many);
+
+bsaved2 := bsaved1 : independent(few);
+
+bsaved3 := bsaved2 : independent(many);
+
+bsaved4 := TABLE(bsaved3, { bsaved3, unsigned forceTransform := 0; }) : independent(few);
+
+bsaved5 := TABLE(bsaved4, { bsaved4 } - [forceTransform]);
+
+output(bsaved5);
